### 2021-07-29 #3551 @whedon
[PRE REVIEW]: Mandyoc: A finite element code to simulate thermochemical convection in parallel
**Submitting author:** @victorsacek (<a href="http://orcid.org/0000-0001-9598-5081">Victor Sacek</a>)
**Repository:** <a href="https://github.com/ggciag/mandyoc" target ="_blank">https://github.com/ggciag/mandyoc</a>
**Version:** v0.1.1
