### 2021-07-29 #3551 @kyleniemeyer
@victorsacek thanks for your interest in JOSS! Our editors in this area are at capacity right now, so I'm going to place this on our waitlist until someone becomes available.
