### 2021-07-31 c78b22c
correction to missplaced note; correction to eqs. weak-formulation-2 and power-law
