### 2021-08-01 cba992b
correction in the script run.sh in the continental rift example
