+++ examples/continental_rift/run.sh
@@ -1,4 +1,4 @@
 /Users/victorsacek/Documents/petsc/arch-label-debug/bin/mpirun -n 8 \
 /Users/victorsacek/Documents/gits/md2d/md2d_aux/mandyoc \
 -seed 0,2 \
--strain_seed 0.0,0.3 \
+-strain_seed 0.0,1.0 \
