### 2021-08-23 #3551 @jedbrown
@aguspesce Great! Just let reply here after making updates. You can comment `@whedon generate pdf` to spin a new PDF at that time.
