### 2021-08-30 ggciag/mandyoc#19 @aguspesce
Fix enviramental variable in src folder
```
src/compile.sh:make all PETSC_DIR=/Users/victorsacek/Documents/petsc PETSC_ARCH=arch-label-debug
src/compile_opt.sh:make all PETSC_DIR=/Users/victorsacek/Documents/petsc PETSC_ARCH=arch-label-optimized
```

https://github.com/openjournals/joss-reviews/issues/3551
