### 2021-08-30 ggciag/mandyoc#21 @aguspesce
Fix warnings during the Mandyoc build
During the Mandyoc  build, there are some warnings ``-Wunused-but-set-variable``. Maybe we can fix.

https://github.com/openjournals/joss-reviews/issues/3551
