### 2021-08-30 ac006c4
Remove compile files from src folder
