+++ src/compile.sh
@@ -1,1 +1,0 @@
-make all PETSC_DIR=/Users/victorsacek/Documents/petsc PETSC_ARCH=arch-label-debug
