### 2021-09-01 8dff657
Remove the initial file in example/continental_rift due to they can be generated using the script.
