+++ examples/continental_rift/param.txt
@@ -1,90 +1,0 @@
-nx = 1601
-nz = 301
-lx = 1600000.0
-lz = 300000.0
-# Simulation options
-multigrid = 1 # ok -> soon to be on the command line only
-solver = direct # default is direct [direct/iterative]
-denok = 1.0E-15 # default is 1.0E-4
-particles_per_element = 40 # default is 81
-particles_perturb_factor = 0.7 # default is 0.5 [values are between 0 and 1]
-rtol = 1.0E-7 # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
-RK4 = Euler # default is Euler [Euler/Runge-Kutta]
-Xi_min = 1.0E-7 # default is 1.0E-14
-random_initial_strain = 0.3 # default is 0.0
-pressure_const = -1.0 # default is -1.0 (not used) - useful only in horizontal 2D models
-initial_dynamic_range = True # default is False [True/False]
-periodic_boundary = False # default is False [True/False]
-high_kappa_in_asthenosphere = False # default is False [True/False]
-K_fluvial = 2.0E-7 # default is 2.0E-7
-m_fluvial = 1.0 # default is 1.0
-sea_level = 0.0 # default is 0.0
-basal_heat = 0.0 # default is -1.0
-# Surface processes
-sp_surface_tracking = False # default is False [True/False]
-sp_surface_processes = False # default is False [True/False]
-sp_dt = 1.0E5 # default is 0.0
-sp_d_c = 1.0 # default is 0.0
-plot_sediment = False # default is False [True/False]
-a2l = True # default is True [True/False]
-free_surface_stab = True # default is True [True/False]
-theta_FSSA = 0.5 # default is 0.5 (only relevant when free_surface_stab = True)
-# Time constrains
-step_max = 5000 # Maximum time-step of the simulation
-time_max = 100.0E6 # Maximum time of the simulation [years]
-dt_max = 10.0E3 # Maximum time between steps of the simulation [years]
-step_print = 10 # Make file every <step_print>
-sub_division_time_step = 0.5 # default is 1.0
-initial_print_step = 0 # default is 0
-initial_print_max_time = 1.0E6 # default is 1.0E6 [years]
-# Viscosity
-viscosity_reference = 1.0E26 # Reference viscosity [Pa.s]
-viscosity_max = 1.0E25 # Maximum viscosity [Pa.s]
-viscosity_min = 1.0E18 # Minimum viscosity [Pa.s]
-viscosity_per_element = constant # default is variable [constant/variable]
-viscosity_mean_method = arithmetic # default is harmonic [harmonic/arithmetic]
-viscosity_dependence = pressure # default is depth [pressure/depth]
-# External ASCII inputs/outputs
-interfaces_from_ascii = True # default is False [True/False]
-n_interfaces = 6 # Number of interfaces int the interfaces.txt file
-variable_bcv = False # default is False [True/False]
-temperature_from_ascii = True # default is False [True/False]
-velocity_from_ascii = True # default is False [True/False]
-binary_output = False # default is False [True/False]
-sticky_blanket_air = True # default is False [True/False]
-precipitation_profile_from_ascii = False # default is False [True/False]
-climate_change_from_ascii = False # default is False [True/False]
-print_step_files = True # default is True [True/False]
-checkered = False # Print one element in the print_step_files (default is False [True/False])
-sp_mode = 5 # default is 1 [0/1/2]
-geoq = on # ok
-geoq_fac = 100.0 # ok
-# Physical parameters
-temperature_difference = 1500. # ok
-thermal_expansion_coefficient = 3.28E-5 # ok
[…]
