+++ examples/continental_rift/interfaces.txt
@@ -1,1608 +1,0 @@
-C 1.0 1.0 0.1 1.0 1.0 1.0 1.0
-rho 3378.0 3354.0 3354.0 3354.0 2800.0 2700.0 1.0
-H 0.0 9.0e-12 9.0e-12 9.0e-12 2.857142857142857e-10 9.25925925925926e-10 0.0
-A 1.393e-14 2.4168E-15 2.4168E-15 2.4168E-15 8.574e-28 8.574e-28 1.0e-18
-n 3.0 3.5 3.5 3.5 4.0 4.0 1.0
-Q 429.0e3 540.0E3 540.0E3 540.0E3 222.0e3 222.0e3 0.0
-V 15.0e-6 25.0e-6 25.0e-6 25.0e-6 0.0 0.0 0.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
--170000.0 -93000.0 -93000.0 -80000.0 -60000.0 -40000.0
[…]
