### 2021-09-01 9eed7c1
Format generate_input_file.py and view_litho_strain_temper.py
