+++ examples/continental_rift/generate_input_files.py
@@ -1,32 +1,26 @@
+"""
+This example simulates the evolution of divergent margins, taking into account the plastic rheology and the sin-rift geodynamics
+"""
+
 import numpy as np
 import matplotlib.pyplot as plt
 
-
 # total model horizontal extent (m)
 Lx = 1600 * 1.0e3
-
 # total model vertical extent (m)
 Lz = 300 * 1.0e3
-
 # number of points in horizontal direction
 Nx = 1601
-
 # number of points in vertical direction
 Nz = 301
-
 # thickness of sticky air layer (m)
 H_sa = 40 * 1.0e3
-
 # thickness of lower crust (m)
 H_lower_crust = 20 * 1.0e3
-
 # thickness of upper crust (m)
 H_upper_crust = 20 * 1.0e3
-
-
 # total thickness of lithosphere (m)
 H_litho = 130 * 1.0e3
-
 # seed depth bellow base of lower crust (m)
 seed_depth = 13 * 1.0e3
 
@@ -39,32 +33,44 @@ X, Z = np.meshgrid(x, z)
 # Interfaces (bottom first)
 ##############################################################################
 interfaces = {
-    'litho': np.ones(Nx) * (H_litho + H_sa),
-    'seed_base': np.ones(Nx) * (seed_depth + H_lower_crust + H_upper_crust + H_sa),
-    'seed_top': np.ones(Nx) * (seed_depth + H_lower_crust + H_upper_crust + H_sa),
-    'lower_crust': np.ones(Nx) * (H_lower_crust + H_upper_crust + H_sa),
-    'upper_crust': np.ones(Nx) * (H_upper_crust + H_sa),
-    'air': np.ones(Nx) * (H_sa),
+    "litho": np.ones(Nx) * (H_litho + H_sa),
+    "seed_base": np.ones(Nx) * (seed_depth + H_lower_crust + H_upper_crust + H_sa),
+    "seed_top": np.ones(Nx) * (seed_depth + H_lower_crust + H_upper_crust + H_sa),
+    "lower_crust": np.ones(Nx) * (H_lower_crust + H_upper_crust + H_sa),
+    "upper_crust": np.ones(Nx) * (H_upper_crust + H_sa),
+    "air": np.ones(Nx) * (H_sa),
 }
 
 # seed thickness (m)
 H_seed = 6 * 1.0e3
-
 # seed horizontal position (m)
 x_seed = 750 * 1.0e3
-
 # seed: number of points of horizontal extent
 n_seed = 2
 
-interfaces['seed_base'][int(Nx*x_seed//Lx - n_seed//2):int(Nx*x_seed//Lx + n_seed//2)] = interfaces['seed_base'][int(Nx*x_seed//Lx - n_seed//2):int(Nx*x_seed//Lx + n_seed//2)] + H_seed//2
-interfaces['seed_top'][int(Nx*x_seed//Lx - n_seed//2):int(Nx*x_seed//Lx + n_seed//2)] = interfaces['seed_top'][int(Nx*x_seed//Lx - n_seed//2):int(Nx*x_seed//Lx + n_seed//2)] - H_seed//2
+interfaces["seed_base"][
+    int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
+] = (
+    interfaces["seed_base"][
+        int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
+    ]
+    + H_seed // 2
+)
+interfaces["seed_top"][
+    int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
+] = (
[…]
