### 2021-09-01 09737f1
Improve the run.sh using env variables
