+++ examples/continental_rift/run.sh
@@ -1,4 +1,6 @@
-/Users/victorsacek/Documents/petsc/arch-label-debug/bin/mpirun -n 8 \
-/Users/victorsacek/Documents/gits/md2d/md2d_aux/mandyoc \
--seed 0,2 \
--strain_seed 0.0,1.0 \
+#!/bin/sh
+
+MANDYOC_PATH = ../..
+MPI_PATH=$HOME/petsc/arch-label-optimized/bin/
+
+MPI_PATH/mpirun -n 8 $MANDYOC_PATH/mandyoc -seed 0,2 -strain_seed 0.0,1.0 
