### 2021-09-01 41c8e1d
Improve the README.
Include a section with instructions for running the example and plotting the results.
