### 2021-09-01 aadf5db
Fix superscript in example/continental_rift/README
