### 2021-09-03 34b7f9e
Remove comment in examples/continental_rift/generate_input_files.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
