### 2021-09-03 1b6099d
Fix typo in examples/continental_rift/generate_input_files.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
