### 2021-09-03 0a9f2fd
Fix typo in examples/continental_rift/generate_input_files.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
