+++ examples/continental_rift/generate_input_files.py
@@ -76,7 +76,7 @@ with open("interfaces.txt", "w") as f:
         H   0.0       9.0e-12    9.0e-12    9.0e-12    {Hlc}       {Huc}       0.0
         A   1.393e-14 2.4168e-15 2.4168e-15 2.4168e-15 8.574e-28   8.574e-28   1.0e-18
         n   3.0       3.5        3.5        3.5        4.0         4.0         1.0
-        Q   429.0e3   540.0E3    540.0E3    540.0E3    222.0e3     222.0e3     0.0
+        Q   429.0e3   540.0e3    540.0e3    540.0e3    222.0e3     222.0e3     0.0
         V   15.0e-6   25.0e-6    25.0e-6    25.0e-6    0.0         0.0         0.0
     """
 
