### 2021-09-03 8fa5b3e
Fix comment translation in examples/continental_rift/generate_input_files.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
