+++ examples/continental_rift/generate_input_files.py
@@ -329,7 +329,7 @@ fac_air = 10.0e3
 vL = 0.005 / (365 * 24 * 3600)  # m/s
 
 h_v_const = 150.0e3  # thickness with constant velocity
-ha = Lz - H_sa - h_v_const  # diferenca
+ha = Lz - H_sa - h_v_const  # difference
 
 vR = 2 * vL * (h_v_const + ha) / ha  # garante que integral seja zero
 
