### 2021-09-03 52a5f85
Fix comment translation in examples/continental_rift/generate_input_files.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
