+++ examples/continental_rift/generate_input_files.py
@@ -365,7 +365,6 @@ cond = Z == 0
 VZ[cond] = vv
 
 # print(np.sum(v0))
-# print(np.sum(vf))
 # print(np.sum(vf)-np.sum(v0))
 
 VVX = np.copy(np.reshape(VX, Nx * Nz))
