### 2021-09-03 1816bbf
Improve README in examples/Crameri2012_case2
* Change superindex.
* Add how to run and plot the results
