### 2021-09-03 8d0bf61
Reformat the python script in examples/Carmeri2012_case2 using Black
