+++ examples/Crameri2012_case2/plot_max_surface.py
@@ -7,13 +7,13 @@ import sys
 import glob
 
 
-param_filename = 'param_1.5.3_2D.txt'
-filename_prefix = 'step_'
-interfaces_filename = 'interfaces_creep.txt'
+param_filename = "param_1.5.3_2D.txt"
+filename_prefix = "step_"
+interfaces_filename = "interfaces_creep.txt"
 
-surface_filenames = sorted(glob.glob(f'sp_surface_global_*'))
+surface_filenames = sorted(glob.glob(f"sp_surface_global_*"))
 
-nproc = len(glob.glob('step_0-rank_new*'))
+nproc = len(glob.glob("step_0-rank_new*"))
 
 s = np.array([])
 t = np.array([])
@@ -22,67 +22,66 @@ size = len(surface_filenames)
 progress = 0
 
 for f in surface_filenames:
-    print(f'{100*progress/(size-1):3.0f}%', end='\r')
+    print(f"{100*progress/(size-1):3.0f}%", end="\r")
     progress += 1
 
+    step = int(f.split(".")[0].split("_")[-1])
 
-    step = int(f.split('.')[0].split('_')[-1])
-
-    _s = np.loadtxt(f, skiprows=2, comments='P')
+    _s = np.loadtxt(f, skiprows=2, comments="P")
     s = np.append(s, np.abs(np.max(_s)))
 
     if step == 0 and int(s) == 0:
         s = 1.5e5
 
     # read time
-    with open(f'time_{step}.txt') as time_file:
+    with open(f"time_{step}.txt") as time_file:
         time_info = time_file.readline()
-        _t = float(time_info.split(':')[-1])
+        _t = float(time_info.split(":")[-1])
         t = np.append(t, _t)
 
 fig, ax = plt.subplots(figsize=(12, 6))
 
 s = 1.5e5 - s
 
-ax.plot(t/1.0e6, s, linestyle='-', color='C0', label='MANDYOC')
+ax.plot(t / 1.0e6, s, linestyle="-", color="C0", label="MANDYOC")
 
-x_underworld, y_underworld = np.loadtxt('UNDERWORLD.txt', unpack=True, delimiter = ',')
-ax.plot(x_underworld, y_underworld, color='C1', label='UNDERWORLD')
+x_underworld, y_underworld = np.loadtxt("UNDERWORLD.txt", unpack=True, delimiter=",")
+ax.plot(x_underworld, y_underworld, color="C1", label="UNDERWORLD")
 
-x_stagyy, y_stagyy = np.loadtxt('STAGYY.txt', unpack=True, delimiter = ',')
-ax.plot(x_stagyy, y_stagyy, color='C2', label='STAGYY')
+x_stagyy, y_stagyy = np.loadtxt("STAGYY.txt", unpack=True, delimiter=",")
+ax.plot(x_stagyy, y_stagyy, color="C2", label="STAGYY")
 
-x_i2ves, y_i2ves = np.loadtxt('I2VES.txt', unpack=True, delimiter = ',')
-ax.plot(x_i2ves, y_i2ves, color='C4', label='I2VIS')
+x_i2ves, y_i2ves = np.loadtxt("I2VES.txt", unpack=True, delimiter=",")
+ax.plot(x_i2ves, y_i2ves, color="C4", label="I2VIS")
 
-plt.legend(loc='upper left')
-ax.grid('on')
+plt.legend(loc="upper left")
+ax.grid("on")
 
 ax.set_xlim((0, 20))
 ax.set_ylim((0, 900))
 
-ax.set_xlabel('Time (Myr)')
-ax.set_ylabel('Maximum topography (m)')
[…]

[…]
