### 2021-09-03 6179acb
Add warning abut the number of cores in example/continental_rift/README
