+++ examples/continental_rift/README.md
@@ -33,10 +33,12 @@ python generate_input_files.py
 
 Now, you can run the model as:
 ```
-~/petsc/arch-label-optimized/bin/mpirun -n 8 ../../mandyoc -seed 0,2 -strain_seed 0.0,1.0 
+~/petsc/arch-label-optimized/bin/mpirun -n number_of_cores ../../mandyoc -seed 0,2 -strain_seed 0.0,1.0 
 ```
+__You have to change `number_of_cores`.__
 
-Or you can use the script called `run.sh` as: 
+
+Or you can use the script called `run.sh` to run it using 8 cores:
 ```
 sh run.sh
 ```
