### 2021-09-03 79416fe
Add same comment in the python script for examples/Crameri2012_case2
