+++ examples/Crameri2012_case2/plot_max_surface.py
@@ -1,3 +1,6 @@
+"""
+Plot the comparison between Mandyoc results and UNDERWORLD, I2VEA and STAGE results
+"""
 import numpy as np
 import numpy.ma as ma
 import matplotlib.pyplot as plt
