+++ examples/Crameri2012_case2/plot_viscosity.py
@@ -1,3 +1,6 @@
+"""
+Plot the viscosity results
+"""
 import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
