### 2021-09-03 29f2a78
Improve examples/Crameri2012_case2/run.sh
