### 2021-09-03 142cd4a
Fix run.sh in examples/vanKeken1997_case1a/run.sh using env variables
