### 2021-09-03 6094afc
Reformat view_density.py in examples/vanKeken1997_case1a
