### 2021-09-06 ggciag/mandyoc#21 @rafaelmds
I got only a few variables marked `Wunused-but-set-variable`

I'll fix those.

> src/DM1_v.cpp: In function ‘PetscErrorCode AssembleA_Veloc(Mat, Mat, DM, DM)’:
> src/DM1_v.cpp:226:12: warning: variable ‘visc_meio’ set but not used [-Wunused-but-set-variable]
>   226 |  PetscReal visc_meio;
>       |            ^~~~~~~~~
> /opt/petsc/debug-v3.14/bin/mpicc -Wall -fdiagnostics-color -c src/DMT.cpp -o src/DMT.o -I/opt/petsc//include -I/opt/petsc//debug-v3.14/include
> /opt/petsc/debug-v3.14/bin/mpicc -Wall -fdiagnostics-color -c src/DMv.cpp -o src/DMv.o -I/opt/petsc//include -I/opt/petsc//debug-v3.14/include
> src/DMv.cpp: In function ‘PetscReal montaKeVeloc_simplif(PetscReal*, PetscReal*, PetscReal*)’:
> src/DMv.cpp:744:9: warning: variable ‘Hx’ set but not used [-Wunused-but-set-variable]
>   744 |  double Hx,Hz;
>       |         ^~
> src/DMv.cpp:744:12: warning: variable ‘Hz’ set but not used [-Wunused-but-set-variable]
>   744 |  double Hx,Hz;
>       |            ^~
