### 2021-09-06 9321413
Update the script to generate the input file for van Keken example
* Now, It generates the parameter file.
