+++ examples/vanKeken1997_case1a/create_interface_file.py
@@ -1,3 +1,6 @@
+"""
+Create the input file to simulate the an Keken et al. (1997) model
+"""
 import numpy as np
 import matplotlib.pyplot as plt
 
@@ -7,27 +10,158 @@ Nz = 81
 Lx = 92.42
 Lz = 100.0
 
-xn = np.linspace(0,Lx,Nx)
+xn = np.linspace(0, Lx, Nx)
 
-Liton = -0.8*Lz + 0.02*Lz*np.cos(np.pi*xn/Lx)
+liton = -0.8 * Lz + 0.02 * Lz * np.cos(np.pi * xn / Lx)
 
-Liton = Liton*1000
-
-f = open("interfaces_creep.txt","w")
-
-f.write("C 1.0 1.0\n")
-f.write("rho -1000. 0.\n")
-f.write("H 0.0E-12 0.0E-12\n")
-f.write("A 0.0 0.0\n")
-f.write("n 0.0 0.0\n")
-f.write("Q 0.0 0.0\n")
-f.write("V 0.0 0.0\n")
-
-for i in np.arange(Nx):
-	f.write("%lf\n"%(Liton[i]))
-
-f.close()
+liton = liton * 1000
 
 
+# Create the interface file
+
+layer_properties = [
+    "C    1.0     1.0",
+    "rho -1000.   0.",
+    "H    0.0E-12 0.0E-12",
+    "A    0.0     0.0",
+    "n    0.0     0.0",
+    "Q    0.0     0.0",
+    "V    0.0     0.0",
+]
+
+with open("interfaces.txt", "w") as f:
+    for line in layer_properties:
+        line = line.strip()
+        if len(line):
+            f.write(" ".join(line.split()) + "\n")
+
+    for line in np.arange(Nx):
+        f.write("%lf\n" % (liton[line]))
 
 
+##############################################################################
+# Parameters file
+##############################################################################
+params = f"""
+# Geometry
+nx                                  = {Nx}          # N. of elements in the longitudinal direction
+nz                                  = {Nz}          # N. of elements in the vertical direction
+lx                                  = 91420.0       # Extent in the longitudinal direction
+lz                                  = 100000.       # Extent in the vertical direction
+particles_per_element_x             = 0             # default is 0
+particles_per_element_z             = 0             # default is 0
+
+# Simulation options
+multigrid                           = 1             # ok -> soon to be on the command line only
+solver                              = direct        # default is direct [direct/iterative]
+denok                               = 1.0E-15       # default is 1.0E-4
+particles_per_element               = 1100          # default is 81
+particles_perturb_factor            = 0.0           # default is 0.5 [values are between 0 and 1]
+rtol                                = 1.0E-5        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
+RK4                                 = Euler         # default is Euler [Euler/Runge-Kutta]
+Xi_min                              = 1.0E-14       # default is 1.0E-14
+random_initial_strain               = 0.0           # default is 0.0
+pressure_const                      = -1.0          # default is -1.0 (not used)
+initial_dynamic_range               = False         # default is False [True/False]
+periodic_boundary                   = False         # default is False [True/False]
+high_kappa_in_asthenosphere         = False         # default is False [True/False]
+K_fluvial                           = 2.0E-7        # default is 2.0E-7
+m_fluvial                           = 1.0           # default is 1.0
+sea_level                           = 0.0           # default is 0.0
+basal_heat                          = -1.0          # default is -1.0
+
+# Surface processes
+sp_surface_tracking                 = False         # default is False [True/False]
+sp_surface_processes                = False         # default is False [True/False]
+sp_dt                               = 0             # default is 0
+sp_d_c                              = 0             # default is 0
[…]
