### 2021-09-06 82029b3
REmove old_version folder for the examples
