+++ examples/old_version/vanKeken1997/param_1.5.3_2D.txt
@@ -1,65 +1,0 @@
-81 81
-91420.0 100000.
-mg 1
-
-stepMAX 4000
-timeMAX 140000.0E6
-
-dt_MAX	1000.0e6
-
-print_step 20
-
-visc 		1.0E21
-visc_MAX	1.0E25
-visc_MIN	1.0E17
-
-n_interfaces	1
-
-geoq_on		1
-geoq_fac 	1.0
-veloc 		0.0E-2
-
-deltaT 		0.
-alpha_exp_thermo 3.28E-5
-kappa 		1.0E-6
-gravity		10.0
-rhom		3300. 
-
-H_per_mass	0.0E-12
-c_heat_capacity 1250.
-
-non_linear	0
-adiabatic_H	0
-radiogenic_H	0
-
-bcv_top_normal	1
-bcv_top_slip	1
-
-bcv_bot_normal	1
-bcv_bot_slip	1
-
-bcv_left_normal	1
-bcv_left_slip	0
-
-bcv_right_normal	1
-bcv_right_slip		0
-
-
-bcT_top		1
-
-bcT_bot		1
-
-bcT_left	0
-
-bcT_right	0
-
-rheol	0
-T_initial	0
-
-H_lito		80000.0
-
-h_air		0.0
-
-beta_max	3.0
-ramp_begin	2000.0E3
-ramp_end	2200.0E3
