+++ examples/old_version/Crameri2012_Case2/I2VES.txt
@@ -1,39 +1,0 @@
-.07597765363128484, 200
-0.09608938547486034, 205.8252427184466
-0.1385474860335194, 219.4174757281553
-0.252513966480447, 240.77669902912623
-0.3217877094972066, 248.54368932038835
-0.39553072625698316, 254.36893203883494
-0.4692737430167595, 260.19417475728153
-0.547486033519553, 266.0194174757281
-0.6458100558659217, 271.8446601941747
-0.7731843575418993, 279.6116504854368
-1.0783055198973044, 295.7446808510639
-1.5147625160462133, 317.0212765957448
-1.8228498074454424, 331.91489361702133
-2.1309370988446723, 348.936170212766
-2.439024390243903, 363.82978723404256
-2.8241335044929397, 385.10638297872345
-3.2862644415917845, 408.51063829787233
-3.6970474967907574, 431.91489361702133
-4.107830551989731, 453.1914893617022
-4.544287548138639, 476.5957446808511
-4.980744544287548, 500.00000000000006
-5.391527599486522, 521.2765957446809
-5.802310654685495, 544.6808510638298
-6.315789473684211, 570.2127659574469
-7.188703465982028, 614.8936170212767
-7.676508344030809, 638.2978723404256
-9.70474967907574, 725.5319148936171
-11.193838254172015, 772.340425531915
-12.297817715019256, 797.872340425532
-13.170731707317072, 812.7659574468087
-14.146341463414634, 825.5319148936171
-14.711168164313221, 831.9148936170213
-15.71245186136072, 838.2978723404257
-16.482670089858793, 842.5531914893618
-17.201540436456995, 842.5531914893618
-18.023106546854944, 842.5531914893618
-18.6392811296534, 840.4255319148938
-19.6405648267009, 836.1702127659576
-19.974326059050064, 834.0425531914894

[…]
