### 2021-09-07 f0e3f74
Create a test which compare the mandyoc output files with an expected result for Crameri model
