### 2021-09-07 c1a996d
Improve Makefile
* Add help
* Add the option to run the test
