+++ AUTHORS.md
@@ -2,10 +2,12 @@
 
 MANDYOC is being developed and currently maintained by
 
-- [Victor Sacek](https://bitbucket.org/victorsacek): mail -
-IAG, Universidad de São Paulo, Brazil - (ORCID: number)
+- [Victor Sacek](https://github.com/victorsacek): sacek@usp.br -
+IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0001-9598-5081)
 
 The following people have made significantly contributions to the project
 (in alphabetical order by last name) are considered "The MANDYOC Developers":
 
-- [developers_name](web): mail - Affiliation - (ORCID: number)
+- [Jamison Assunção](https://github.com/jamisonassuncao): mail - IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0003-2822-2417)
+- [Agustina Pesce](https://github.com/aguspesce): mail - Instituto Geofísico Sismológico Ing. Volponi, Universidad Nacional de San Juan, Argentina; CONICET, Argentina - (ORCID: 0000-0002-5538-8845)
+- [Rafael Monteiro da Silva](https://github.com/rafaelmds): rafael.m.silva@alumni.usp.br - IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0001-8645-2443)
