### 2021-09-09 252ded9
Update the petsc path in  Makefile
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
