### 2021-09-09 9bdf62a
Update test path in Makefile
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
