### 2021-09-09 fc8732c
Update testing data paht in test/testing_result.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
