### 2021-09-09 5e8d177
Update the expected data path in test/testing_result.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
