+++ test/testing_result.py
@@ -8,7 +8,7 @@ import os
 # Test path
 test_path = "test/testing_data"
 # Path to the expected data to make the comparison
-expected_result_path = os.path.join(test_path, "expected_result")
+expected_result_path = os.path.join(test_path, "expected")
 # Name of the files to compare
 file_name = [
     "density_{}", 
