### 2021-09-09 81e8ebe
Change the number of steps to compare in test / testing_result.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
