### 2021-09-09 0948e82
Fix spaces in  test/testing_result.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
