### 2021-09-09 62c4a19
Fix spaces in test/testing_result.py
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
