### 2021-09-09 3a114e2
Improve docstring
