### 2021-09-09 2e5a3d6
Modify testing_result.py to use pytest
