### 2021-09-09 1283d13
Update the Makefile with pytest
