### 2021-09-10 192bbd7
Fix  scientific notation in examples/vanKeken1997_case1a/create_interface_file.py
* Change E to lower case

Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
