### 2021-09-10 510e1a2
Add number of code as env variable in  examples/Crameri2012_case2/run.sh
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
