### 2021-09-10 2516b76
Put number of cores as env variable in  examples/continental_rift/run.sh
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
