+++ examples/vanKeken1997_case1a/create_interface_file.py
@@ -22,7 +22,7 @@ liton = liton * 1000
 layer_properties = [
     "C    1.0     1.0",
     "rho -1000.   0.",
-    "H    0.0E-12 0.0E-12",
+    "H    0.0e-12 0.0e-12",
     "A    0.0     0.0",
     "n    0.0     0.0",
     "Q    0.0     0.0",
