### 2021-09-10 83d5f05
Add number of cores as env variable in examples/vanKeken1997_case1a/run.sh
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
