### 2021-09-10 cff05c3
Add petsc path as env variable in  examples/Crameri2012_case2/README.md
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
