### 2021-09-10 59adc42
Add petsc as env variable in  examples/continental_rift/README.md
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
