### 2021-09-10 35c8082
Update examples/continental_rift/README.md
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
