### 2021-09-10 35b843a
add action to build docker image
