### 2021-09-10 3bbc267
Fix the mpi path in all run.sh for the examples
