+++ examples/Crameri2012_case2/run.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-MANDYOC_PATH = ../..
-MPI_PATH=${HOME}/${PETSC_DIR}/${PETSC_ARCH}/bin/
+MPI_PATH=$HOME/petsc/arch-label-optimized/bin
+MANDYOC_PATH=../..
 NUMBER_OF_CORES=2
 
-MPI_PATH/mpirun -n $NUMBER_OF_CORES $MANDYOC_PATH/mandyoc
+$MPI_PATH/mpirun -n $NUMBER_OF_CORES $MANDYOC_PATH/mandyoc
