+++ examples/vanKeken1997_case1a/run.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-MANDYOC_PATH = ../..
-MPI_PATH=${HOME}/${PETSC_DIR}/${PETSC_PATH}/bin/
+MANDYOC_PATH =../..
+MPI_PATH=$HOME/${PETSC_DIR}/${PETSC_PATH}/bin/
 NUMBER_OF_CORES=2
 
 MPI_PATH/mpirun -n $NUMBER_OF_CORES $MANDYOC_PATH/mandyoc
