### 2021-09-10 9e9cd13
Fix petsc path in vanKeken example
