+++ examples/Crameri2012_case2/generate_input_files.py
@@ -109,20 +109,20 @@ lz                                  = {Lz:.1e}       # Extent in the vertical di
 # Simulation options
 multigrid                           = 1
 solver                              = direct        # default is direct [direct/iterative]
-denok                               = 1.0E-15       # default is 1.0E-4
+denok                               = 1.0e-15       # default is 1.0E-4
 particles_per_element               = 600           # default is 81
 particles_per_element_x             = 6             # Number of particles per element in longitudinal (default is 0)
 particles_per_element_z             = 100           # Number of particles per element in vertical (default is 0)
 particles_perturb_factor            = 0.7           # default is 0.5 [values are between 0 and 1]
-rtol                                = 1.0E-7        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
+rtol                                = 1.0e-7        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
 RK4                                 = Euler         # default is Euler [Euler/Runge-Kutta]
-Xi_min                              = 1.0E-14       # default is 1.0E-14
+Xi_min                              = 1.0e-14       # default is 1.0E-14
 random_initial_strain               = 0.0           # default is 0.0
 pressure_const                      = -1.0          # default is -1.0 (not used) - useful only in horizontal 2D models
 initial_dynamic_range               = False         # default is False [True/False]
 periodic_boundary                   = False         # default is False [True/False]
 high_kappa_in_asthenosphere         = False         # default is False [True/False]
-K_fluvial                           = 2.0E-7        # default is 2.0E-7
+K_fluvial                           = 2.0e-7        # default is 2.0E-7
 m_fluvial                           = 1.0           # default is 1.0
 sea_level                           = 0.0           # default is 0.0
 basal_heat                          = -1.0          # default is -1.0
@@ -141,17 +141,17 @@ theta_FSSA                          = 0.5           # default is 0.5 (only relev
 
 # Time constrains
 step_max                            = 7000          # Maximum time-step of the simulation
-time_max                            = 21.0E6        # Maximum time of the simulation [years]
-dt_max                              = 10.0E3        # Maximum time between steps of the simulation [years]
+time_max                            = 21.0e6        # Maximum time of the simulation [years]
+dt_max                              = 10.0e3        # Maximum time between steps of the simulation [years]
 step_print                          = 10            # Make file every <step_print>
 sub_division_time_step              = 1.0           # default is 1.0
 initial_print_step                  = 2             # default is 0
-initial_print_max_time              = 1.0E6         # default is 1.0E6 [years]
+initial_print_max_time              = 1.0e6         # default is 1.0e6 [years]
 
 # Viscosity
-viscosity_reference                 = 1.0E21        # Reference viscosity [Pa.s]
-viscosity_max                       = 1.0E23        # Maximum viscosity [Pa.s]
-viscosity_min                       = 1.0E19        # Minimum viscosity [Pa.s]
+viscosity_reference                 = 1.0e21        # Reference viscosity [Pa.s]
+viscosity_max                       = 1.0e23        # Maximum viscosity [Pa.s]
+viscosity_min                       = 1.0e19        # Minimum viscosity [Pa.s]
 viscosity_per_element               = constant      # default is variable [constant/variable]
 viscosity_mean_method               = arithmetic    # default is harmonic [harmonic/arithmetic]
 viscosity_dependence                = pressure      # default is depth [pressure/depth]
@@ -175,11 +175,11 @@ geoq_fac                            = 100.0
 
 # Physical parameters
 temperature_difference              = 0.0
-thermal_expansion_coefficient       = 3.28E-5
[…]
