### 2021-09-14 1b00a78
Fix typo in  examples/Crameri2012_case2/README.md
Co-authored-by: victorsacek <sacek@usp.br>
