### 2021-09-14 f3d3e4f
Add the dependencies in examples README files
