+++ examples/vanKeken1997_case1a/README.md
@@ -9,6 +9,12 @@ under a denser thicker package.
 
 ## How to run this example
 
+### Dependencies
+
+To run this example you need to install:
+
+* `numpy`
+
 ### Generate the input files
 
 To generate the interface, velocity, temperature and parameter file, you need to run `generate_interface.py` as:
