### 2021-09-14 0c77ab9
Add how to test the code in the CONTRIBUTING guidelines file
