### 2021-09-14 d20f674
Add  numpy adn pytest as dependencies in documentation
