### 2021-09-14 4c10e87
move "how to run examples" to examples section
Co-authored-by: Agustina <pesce.agustina@gmail.com>
