### 2021-09-14 13e75d9
update sections examples to add instructions on how to run examples
