### 2021-09-16 d92d312
Update developers dependencies
Ass numpy and os
