### 2021-09-16 03b3bd3
Improve examples/vanKeken1997_case1a/view_density.py
Fix how the script reads the parameter file
