+++ examples/vanKeken1997_case1a/view_density.py
@@ -1,13 +1,21 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-with open("param_1.5.3_2D.txt", "r") as f:
+with open("param.txt", "r") as f:
+    line = f.readline()
+    line = f.readline()
     line = f.readline()
     line = line.split()
-    Nx, Nz = int(line[0]), int(line[1])
+    Nx = int(line[2])
     line = f.readline()
     line = line.split()
-    Lx, Lz = float(line[0]), float(line[1])
+    Nz = int(line[2])
+    line = f.readline()
+    line = line.split()
+    Lx = float(line[2])
+    line = f.readline()
+    line = line.split()
+    Lz = float(line[2])
 
 print(Nx, Nz, Lx, Lz)
 
@@ -21,19 +29,19 @@ for cont in range(0, 4000, 40):
     print(cont)
     try:
         if n_cores > 1:
-            tempo = np.loadtxt(
-                "Tempo_" + str(cont) + ".txt", unpack=True, delimiter=":", usecols=(1)
+            time = np.loadtxt(
+                "time_" + str(cont) + ".txt", unpack=True, delimiter=":", usecols=(1)
             )
-            tempo = tempo[0]
+            time = time[0]
         else:
-            tempo = np.loadtxt(
-                "Tempo_" + str(cont) + ".txt", unpack=True, delimiter=":"
+            time = np.loadtxt(
+                "time_" + str(cont) + ".txt", unpack=True, delimiter=":"
             )
     except:
         print("Step %d not found" % (cont))
         break
 
-    A = np.loadtxt("Rho_" + str(cont) + ".txt", unpack=True, comments="P", skiprows=2)
+    A = np.loadtxt("density_" + str(cont) + ".txt", unpack=True, comments="P", skiprows=2)
     TT = A * 1.0
     TT[np.abs(TT) < 1.0e-200] = 0
     TT = np.reshape(TT, (Nx, Nz), order="F")
@@ -41,9 +49,6 @@ for cont in range(0, 4000, 40):
     plt.close()
     plt.figure(figsize=(2, 2))
     plt.imshow(np.transpose(TTT[:, ::-1]), extent=[0, 0.9142, 0, 1.0])
-    plt.title("t' = %.1f" % (tempo * 365 * 24 * 3600.0 * 1.0e-12))  # adimensional time
-    # tempo in years
-    # t' = tempo * seconds_per_year * kappa / h^2
-    # t' = tempo * seconds_per_year * (rho g h^3/ Rb visc) / h^2
+    plt.title("t' = %.1f" % (time * 365 * 24 * 3600.0 * 1.0e-12))  # adimensional time
 
-    plt.savefig("Rho_{:05}.png".format(cont))
+    plt.savefig("density_{:05}.png".format(cont))
