### 2021-09-16 209c402
Update the Makefile.
Now you can change the number of cores to run the test.
As default it is equal to 2.
