### 2021-09-16 1954dd3
Revert "Update developers dependencies"
This reverts commit d92d312ca18c58cc7a515a96a9d536ed80f51253.
