+++ docs/files/installation.rst
@@ -24,6 +24,9 @@ Dependencies
 **Developers dependencies**
 
 * pytest
+* numpy
+* os
+
 
 PETSc Installation
 ------------------
