### 2021-09-17 ab3de42
Rename the output file called sp_surface_global_*.txt
Eliminate the number of zeros before the dot
