+++ AUTHORS.md
@@ -8,6 +8,6 @@ IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0001-9598-5081)
 The following people have made significantly contributions to the project
 (in alphabetical order by last name) are considered "The MANDYOC Developers":
 
-- [Jamison Assunção](https://github.com/jamisonassuncao): mail - IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0003-2822-2417)
+- [Jamison Assunção](https://github.com/jamisonassuncao): jamison.assuncao@usp.br - IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0003-2822-2417)
 - [Agustina Pesce](https://github.com/aguspesce): pesce.agustina@gmail.com - Instituto Geofísico Sismológico Ing. Volponi, Universidad Nacional de San Juan, Argentina; CONICET, Argentina - (ORCID: 0000-0002-5538-8845)
 - [Rafael Monteiro da Silva](https://github.com/rafaelmds): rafael.m.silva@alumni.usp.br - IAG, Universidade de São Paulo, Brazil - (ORCID: 0000-0001-8645-2443)
