### 2021-09-20 50883a3
Remove variables that were not used in exampeles/Crameri2012_case2/plot_max_sueface.py
