### 2021-09-20 5ffc9ba
Fix typo in Makefile
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
