### 2021-09-20 93965e4
Add the sp_surface_global in the test script
