### 2021-09-20 b36ca9a
add path constrain to trigger build-docs-web action only for changes in *docs* folder (#33)
