### 2021-09-27 b40309f
Improve examples  (#25)
* Remove the initial file in example/continental_rift due to
they can be generated using the script.

* Add the initial files for the examples in the gitignore

* Remove figures because they can be generated using the script

* Add the example/continental_rift/*.png in gitignore

* Fix typo on gitignore

* Format generate_input_file.py and view_litho_strain_temper.py

* Improve the run.sh using env variables

* Improve the README.

Include a section with instructions for running the example and plotting the results.

* Add subtitle in example/continental_rift/README

* 
[…]
