### 2021-09-27 740900a
add information about examples (#30)
* add information about examples

* move "how to run examples" to examples section

Co-authored-by: Agustina <pesce.agustina@gmail.com>

* update sections examples to add instructions on how to run examples

Co-authored-by: Agustina <pesce.agustina@gmail.com>
