### 2021-09-27 bbe069c
Add mising dependencies to run the examples
