### 2021-09-27 b2b969b
Remove interfaces file from vanKeken example
