+++ examples/vanKeken1997_case1a/interfaces.txt
@@ -1,88 +1,0 @@
-C 1.0 1.0
-rho -1000. 0.
-H 0.0E-12 0.0E-12
-A 0.0 0.0
-n 0.0 0.0
-Q 0.0 0.0
-V 0.0 0.0
--78000.000000
--78001.541928
--78006.165333
--78013.863086
--78024.623319
--78038.429439
--78055.260159
--78075.089527
--78097.886967
--78123.617328
--78152.240935
--78183.713652
--78217.986952
--78255.007986
--78294.719671
--78337.060775
--78381.966011
--78429.366138
--78479.188069
--78531.354981
--78585.786438
--78642.398509
--78701.103903
--78761.812101
--78824.429495
--78888.859534
--78955.002871
--79022.757517
--79092.019001
--79162.680525
--79234.633135
--79307.765886
--79381.966011
--79457.119100
--79533.109272
--79609.819356
--79687.131070
--79764.925205
--79843.081809
--79921.480368
--80000.000000
--80078.519632
--80156.918191
--80235.074795
--80312.868930
--80390.180644
--80466.890728
--80542.880900
--80618.033989
--80692.234114
--80765.366865
--80837.319475
--80907.980999
--80977.242483
--81044.997129
--81111.140466
--81175.570505
--81238.187899
--81298.896097
--81357.601491
--81414.213562
--81468.645019
--81520.811931
--81570.633862
--81618.033989
--81662.939225
--81705.280329
--81744.992014
--81782.013048
--81816.286348
--81847.759065
--81876.382672
--81902.113033
--81924.910473
--81944.739841
--81961.570561
--81975.376681
--81986.136914
--81993.834667
--81998.458072
--82000.000000
