### 2021-09-27 fbb7e84
Put how run the Mnadyoc test in the documentation
