+++ docs/files/installation.rst
@@ -19,14 +19,9 @@ Dependencies
 **Optional** dependencies:
 
 * gfortran
-* numpy to run the examples.
-
-**Developers dependencies**
-
 * pytest
 * numpy
 * os
-
 
 PETSc Installation
 ------------------
@@ -107,6 +102,12 @@ Next, **build and install** *Mandyoc* by running::
 
 	To print *Mandyoc* runtime options, run mandyoc with `-flags` command line
 	argument.
+
+**Check** Mandyoc installation with:
+
+.. code-block::
+
+	make test_mandyoc
 
 Examples
 ---------------
