### 2021-09-27 852afff
Update paper/paper.md
Co-authored-by: victorsacek <sacek@usp.br>
