### 2021-09-27 bcc7102
Added info in the paper about temporat vbc and surface processes (#34)
* added info in the paper about temporat vbc and surface processes

* Update paper/paper.md

Co-authored-by: victorsacek <sacek@usp.br>

Co-authored-by: victorsacek <sacek@usp.br>
