### 2021-09-27 fbfb361
Add mising dependencies to run the examples (#35)
* Add mising dependencies to run the examples

* Remove interfaces file from vanKeken example
