### 2021-09-27 29602c2
Test to check that the Mandyoc output result for a given model is equal to the expected result (#27)
* Create a test which compare the mandyoc output files with an
expected result for Crameri model

* Add mandyoc output file in gitignore

* Improve Makefile

* Add help
* Add the option to run the test

* Update .gitignore

Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>

* Update the petsc path in  Makefile

Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>

* Update test path in Makefile

Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>

* Up
[…]
