### 2021-09-27 b194fba
fix expression for viscosity component in docs
