### 2021-09-27 #3551 @victorsacek
Hello @jedbrown,

Thank you very much for the review of our code.

We've made almost all the changes that you suggested:
* Improved the examples (#25):
    * Fixed path.
    * Added README file with running instructions.
    * Improved examples script to create input files and plot results.
* Made test to check if Mandyoc's output result for a given model is equal to the expected result (#27)
* Improved paper (#34)
* Added developers info (#28)
* Fixed -Wunused-but-set-variable warnings during Mandyoc compilation (#26)
* Fixed documentation (#16 #33 #30)
* Improved Github Actions for documentation (#33)

We have to clarify that we decided not to create an automated test using GitHub actions because installing all dependencies to run Mandyoc and test it would take too much time for the CI.
But we created a test based on regression testing.
This test can be executed once the program is installed.
We intend to improve the code to use Docker in the future.

All the changes are in the release [v0.1.2](https://github.com/ggciag/mandyoc/releases/tag/v0.1.2).

Please inform us if the present version of the code is ok for review in JOSS.

Thank you for your attention.

@aguspesce @jamisonassuncao @rafaelmds
