+++ src/reader.cpp
@@ -7,9 +7,17 @@ PetscBool check_a_b_bool(char tkn_w[], char tkn_v[], const char str_a[], const c
 void ErrorInterfaces();
 
 // Parameter file variables
-extern long Nx, Nz;
+extern long T_NE;
+extern long T_GN;
+extern int DIMEN;
+extern long GaussQuad;
+extern long V_NE;
+extern long V_GN;
+extern long V_GT;
+
+extern long Nx, Ny, Nz;
 extern long layers;
-extern double Lx, depth;
+extern double Lx, Ly, depth;
 extern int ContMult;
 extern long stepMAX;
 extern double timeMAX;
@@ -77,6 +85,7 @@ extern PetscReal pressure_const;
 extern PetscInt initial_dynamic_range;
 extern PetscInt periodic_boundary;
 extern PetscInt nx_ppe;
+extern PetscInt ny_ppe;
 extern PetscInt nz_ppe;
 extern PetscInt initial_print_step;
 extern PetscReal initial_print_max_time;
@@ -168,8 +177,10 @@ PetscErrorCode reader(int rank, const char fName[]){
 			// Store and check every parameter.
 			// Parameters with values
 			if (strcmp(tkn_w, "nx") == 0) {Nx = atol(tkn_v);}
+			else if (strcmp(tkn_w, "ny") == 0) {Ny = atol(tkn_v);}
 			else if (strcmp(tkn_w, "nz") == 0) {Nz = atol(tkn_v);}
 			else if (strcmp(tkn_w, "lx") == 0) {Lx = atof(tkn_v);}
+			else if (strcmp(tkn_w, "ly") == 0) {Ly = atof(tkn_v);}
 			else if (strcmp(tkn_w, "lz") == 0) {depth = atof(tkn_v);}
 			else if (strcmp(tkn_w, "multigrid") == 0) {ContMult = atoi(tkn_v);}
 			else if (strcmp(tkn_w, "step_max") == 0) {stepMAX = atol(tkn_v);}
@@ -202,6 +213,7 @@ PetscErrorCode reader(int rank, const char fName[]){
 			else if (strcmp(tkn_w, "random_initial_strain") == 0) {random_initial_strain = atof(tkn_v);}
 			else if (strcmp(tkn_w, "pressure_const") == 0) {pressure_const = atof(tkn_v);}
 			else if (strcmp(tkn_w, "particles_per_element_x") == 0) {nx_ppe = atoi(tkn_v);}
+			else if (strcmp(tkn_w, "particles_per_element_y") == 0) {ny_ppe = atoi(tkn_v);}
 			else if (strcmp(tkn_w, "particles_per_element_z") == 0) {nz_ppe = atoi(tkn_v);}
 			else if (strcmp(tkn_w, "initial_print_step") == 0) {initial_print_step = atoi(tkn_v);}
 			else if (strcmp(tkn_w, "initial_print_max_time") == 0) {initial_print_max_time = atof(tkn_v);}
@@ -270,6 +282,7 @@ PetscErrorCode reader(int rank, const char fName[]){
 		if (particles_perturb_factor < 0.0) particles_perturb_factor = 0.0;
 		layers = Nz;
 		if (nx_ppe < 0) nx_ppe = 0;
[…]
