### 2021-10-11 2c6cc1a
corrections until create_thermal_2d_3d
