+++ src/DMT.cpp
@@ -25,8 +25,11 @@ extern double Lx, Ly, depth;
 
 extern double dz_const;
 
+extern int DIMEN;
+
 extern PetscReal *NT;
 extern PetscReal *NT_x;
+extern PetscReal *NT_y;
 extern PetscReal *NT_z;
 
 extern long GaussQuad;
@@ -94,7 +97,7 @@ extern double kappa;
 extern double RHOM;
 extern double c_heat_capacity;
 
-PetscErrorCode create_thermal_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz)
+PetscErrorCode create_thermal_2d_3d(PetscInt mx,PetscInt my,PetscInt mz,PetscInt Px,PetscInt Py,PetscInt Pz)
 {
 
 	PetscInt       dof,stencil_width;
@@ -110,15 +113,22 @@ PetscErrorCode create_thermal_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz
 	PetscTime(&Tempo1p);
 	
 	dof           = 1;
-	if (periodic_boundary==0)	{
-		stencil_width = 1;
-		ierr          = DMDACreate2d(PETSC_COMM_WORLD,DM_BOUNDARY_NONE,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
-								 mx+1,mz+1,Px,Pz,dof,stencil_width,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
+	if (DIMEN==2){
+		if (periodic_boundary==0)	{
+			stencil_width = 1;
+			ierr          = DMDACreate2d(PETSC_COMM_WORLD,DM_BOUNDARY_NONE,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
+									mx+1,mz+1,Px,Pz,dof,stencil_width,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
+		}
+		else {
+			stencil_width = 2;
+			ierr          = DMDACreate2d(PETSC_COMM_WORLD,DM_BOUNDARY_PERIODIC,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
+									mx+1,mz+1,Px,Pz,dof,stencil_width,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
+		}
 	}
 	else {
-		stencil_width = 2;
-		ierr          = DMDACreate2d(PETSC_COMM_WORLD,DM_BOUNDARY_PERIODIC,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
-								 mx+1,mz+1,Px,Pz,dof,stencil_width,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
+		stencil_width = 1;
+		ierr          = DMDACreate3d(PETSC_COMM_WORLD,DM_BOUNDARY_NONE,DM_BOUNDARY_NONE,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
+								 mx+1,my+1,mz+1,Px,Py,Pz,dof,stencil_width,NULL,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
 	}
 
 	ierr = DMSetFromOptions(da_Thermal);CHKERRQ(ierr);
@@ -131,6 +141,9 @@ PetscErrorCode create_thermal_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz
 	
 	ierr = PetscCalloc1(GaussQuad*T_NE,&NT); CHKERRQ(ierr);
[…]
