### 2021-10-15 5acad3b
update in DM1_v and DMv files for the incorporation of the 3D version
