+++ src/DMv.cpp
@@ -24,8 +24,15 @@ extern PetscInt visc_const_per_element;
 typedef struct {
 	PetscScalar u;
 	PetscScalar w;
-	//PetscScalar p; //check for the future 3d version
+	//PetscScalar p;
 } Stokes;
+
+typedef struct {
+	PetscScalar u;
+	PetscScalar v;
+	PetscScalar w;
+	//PetscScalar p;
+} Stokes3d;
 
 PetscErrorCode ascii2bin(char *s1, char *s2);
 
@@ -33,13 +40,16 @@ PetscErrorCode AssembleA_Veloc(Mat A,Mat AG,DM veloc_da, DM temper_da);
 
 PetscErrorCode AssembleF_Veloc(Vec F,DM veloc_da,DM drho_da, Vec FP);
 
-PetscErrorCode montaKeVeloc_general(PetscReal *KeG, double dx_const, double dz_const);
+PetscErrorCode montaKeVeloc_general_2d(PetscReal *KeG, double dx_const, double dz_const);
+PetscErrorCode montaKeVeloc_general_3d(PetscReal *KeG, double dx_const, double dy_const, double dz_const);
 
-PetscReal montaKeVeloc_simplif(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
+PetscReal montaKeVeloc_simplif_2d(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
 
-PetscErrorCode montaCeVeloc(PetscReal *Ce);
+PetscErrorCode montaCeVeloc_2d(PetscReal *Ce);
+PetscErrorCode montafeVeloc_2d(PetscReal *fMe);
 
-PetscErrorCode montafeVeloc(PetscReal *fMe);
+PetscErrorCode montaCeVeloc_3d(PetscReal *Ce);
+PetscErrorCode montafeVeloc_3d(PetscReal *fMe);
 
 PetscErrorCode calc_drho();
 
@@ -64,6 +74,8 @@ PetscErrorCode Swarm2Mesh();
 extern double r06;
 extern double r8p9;
 extern double r5p9;
+
+extern int DIMEN;
 
 extern long V_NE, V_GN, V_GT;
 
@@ -110,6 +122,7 @@ extern PetscReal *Vfe;
 
 
 extern double dx_const;
+extern double dy_const;
 extern double dz_const;
 
 
@@ -145,7 +158,7 @@ extern int n_interfaces;
 
 extern PetscInt binary_output;
 
-PetscErrorCode create_veloc_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz)
+PetscErrorCode create_veloc_2d_3d(PetscInt mx,PetscInt my,PetscInt mz,PetscInt Px,PetscInt Py,PetscInt Pz)
 {
 
 	PetscInt       dof,stencil_width;
@@ -164,23 +177,36 @@ PetscErrorCode create_veloc_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz)
 
 	PetscFunctionBeginUser;
 
-
-	dof           = 2; //check for the future 3d version
-	if (periodic_boundary==0){
-		stencil_width = 1;
[…]
