+++ src/DMSwarm.cpp
@@ -7,6 +7,8 @@
 #include <petsc/private/dmimpl.h>
 #include <petscmath.h>
 
+extern int DIMEN;
+
 extern DM dmcell;
 
 extern DM dms;
@@ -14,17 +16,20 @@ extern DM dms;
 extern DM da_Veloc;
 
 extern long Nx;
+extern long Ny;
 
 extern double dx_const;
+extern double dy_const;
 extern double dz_const;
 
-extern double Lx, depth;
+extern double Lx, Ly, depth;
 
 extern double H_lito;
 extern double escala_viscosidade;
 
 extern PetscInt particles_per_ele;
 extern PetscInt nx_ppe;
+extern PetscInt ny_ppe;
 extern PetscInt nz_ppe;
 
 extern double H_per_mass;
@@ -268,22 +273,35 @@ PetscErrorCode createSwarm()
 	PetscReal *rarray;
 	PetscReal *strain_array;
 
+	PetscInt nx_part, ny_part, nz_part;
 
-	PetscInt nz_part = (int)PetscSqrtReal(particles_per_ele*dz_const/dx_const);
-	PetscInt nx_part = (int)(particles_per_ele/nz_part);
+	if (DIMEN==2){
+		nz_part = (int)PetscSqrtReal(particles_per_ele*dz_const/dx_const);
+		nx_part = (int)(particles_per_ele/nz_part);
 
-	if (nx_ppe > 0) {
-		nx_part = nx_ppe;
+		if (nx_ppe > 0) {nx_part = nx_ppe;}
+		if (nz_ppe > 0) {nz_part = nz_ppe;}
+
+		particles_per_ele = nx_part*nz_part;
 	}
+	else {
+		nx_part = (int)PetscCbrtReal(particles_per_ele*dx_const*dx_const/(dy_const*dz_const));
+		ny_part = (int)PetscCbrtReal(particles_per_ele*dy_const*dy_const/(dz_const*dx_const));
+		nz_part = (int)PetscCbrtReal(particles_per_ele*dz_const*dz_const/(dx_const*dy_const));
 
-	if (nz_ppe > 0) {
-		nz_part = nz_ppe;
+		if (nx_ppe > 0) {nx_part = nx_ppe;}
+		if (ny_ppe > 0) {ny_part = ny_ppe;}
+		if (nz_ppe > 0) {nz_part = nz_ppe;}
+
+		particles_per_ele = nx_part*ny_part*nz_part;
+		
 	}
-
-	particles_per_ele = nx_part*nz_part;
 
 
 	PetscPrintf(PETSC_COMM_WORLD,"particles per element in x:  %d\n",nx_part);
+	if (DIMEN==3){
+		PetscPrintf(PETSC_COMM_WORLD,"particles per element in y:  %d\n",ny_part);
+	}
 	PetscPrintf(PETSC_COMM_WORLD,"particles per element in z:  %d\n",nz_part);
 	PetscPrintf(PETSC_COMM_WORLD,"total particles per element: %d\n\n",particles_per_ele);
 
@@ -296,7 +314,7 @@ PetscErrorCode createSwarm()
 	/* Create the swarm */
 	ierr = DMCreate(PETSC_COMM_WORLD,&dms);CHKERRQ(ierr);
 	ierr = DMSetType(dms,DMSWARM);CHKERRQ(ierr);
-	ierr = DMSetDimension(dms,2);CHKERRQ(ierr);
+	ierr = DMSetDimension(dms,DIMEN);CHKERRQ(ierr);
 
[…]
