+++ src/DM1.cpp
@@ -10,6 +10,8 @@ PetscErrorCode mean_value_periodic_boundary(DM da,Vec F,Vec local_F, PetscScalar
 extern double alpha_exp_thermo;
 extern double gravity;
 
+extern int DIMEN;
+
 extern PetscReal *TCe;
 extern PetscReal *TCe_fut;
 
@@ -34,7 +36,7 @@ extern Vec Temper_Cond;
 
 extern double Delta_T;
 
-extern double Lx, depth;
+extern double Lx, Ly, depth;
 
 extern double h_air;
 
@@ -68,6 +70,7 @@ double Thermal_profile(double t, double zz);
 
 PetscReal Temper3(double xx,double zz);
 PetscReal Temper4(double xx,double zz);
+PetscReal Temper4_3D(double xx,double yy,double zz);
 
 
 extern PetscInt WITH_NON_LINEAR;
@@ -109,6 +112,31 @@ PetscErrorCode DMDAGetLocalElementSize(DM da,PetscInt *mxl,PetscInt *mzl)
 	PetscFunctionReturn(0);
 }
 
+PetscErrorCode DMDAGetLocalElementSize_3d(DM da,PetscInt *mxl,PetscInt *myl,PetscInt *mzl)
+{
+	PetscInt       m,n,p,M,N,P;
+	PetscInt       sx,sy,sz;
+	PetscErrorCode ierr;
+	
+	PetscFunctionBeginUser;
+	ierr = DMDAGetInfo(da,0,&M,&N,&P,0,0,0,0,0,0,0,0,0);CHKERRQ(ierr);
+	ierr = DMDAGetCorners(da,&sx,&sy,&sz,&m,&n,&p);CHKERRQ(ierr);
+	
+	if (mxl) {
+		*mxl = m;
+		if ((sx+m) == M) *mxl = m-1;  /* last proc */
+	}
+	if (myl) {
+		*myl = n;
+		if ((sy+n) == N) *myl = n-1;  /* last proc */
+	}
+	if (mzl) {
+		*mzl = p;
+		if ((sz+p) == P) *mzl = p-1;  /* last proc */
+	}
+	PetscFunctionReturn(0);
+}
+
 
 PetscErrorCode DMDAGetElementCorners(DM da,PetscInt *sx,PetscInt *sz,PetscInt *mx,PetscInt *mz)
 {
@@ -127,6 +155,30 @@ PetscErrorCode DMDAGetElementCorners(DM da,PetscInt *sx,PetscInt *sz,PetscInt *m
 		if (sk != 0) *sz = sk+1+periodic_boundary;
 	}
 	ierr = DMDAGetLocalElementSize(da,mx,mz);CHKERRQ(ierr);
+	PetscFunctionReturn(0);
+}
+
+PetscErrorCode DMDAGetElementCorners3d(DM da,PetscInt *sx,PetscInt *sy,PetscInt *sz,PetscInt *mx,PetscInt *my,PetscInt *mz)
+{
+	PetscInt       si,sj,sk;
+	PetscErrorCode ierr;
+	
+	PetscFunctionBeginUser;
+	ierr = DMDAGetGhostCorners(da,&si,&sj,&sk,0,0,0);CHKERRQ(ierr);
+	
+	if (sx) {
+		*sx = si;
+		if (si != 0) *sx = si+1;
+	}
+	if (sy) {
+		*sy = sj;
+		if (sj != 0) *sy = sj+1;
+	}
+	if (sz) {
+		*sz = sk;
+		if (sk != 0) *sz = sk+1;
+	}
+	ierr = DMDAGetLocalElementSize_3d(da,mx,my,mz);CHKERRQ(ierr);
 	PetscFunctionReturn(0);
 }
 
[…]
