+++ src/DMSwarm2mesh.cpp
@@ -17,12 +17,13 @@ extern Vec geoq_strain_rate, local_geoq_strain_rate;
 
 extern Vec Temper,local_Temper;
 
-extern long Nx,Nz;
+extern long Nx,Ny,Nz;
 
 extern double dx_const;
+extern double dy_const;
 extern double dz_const;
 
-extern double Lx, depth;
+extern double Lx,Ly, depth;
 
 extern PetscInt visc_harmonic_mean;
 extern PetscInt visc_const_per_element;
@@ -36,7 +37,7 @@ extern PetscInt periodic_boundary;
 
 extern double visc_MIN;
 
-PetscErrorCode Swarm2Mesh(){
+PetscErrorCode Swarm2Mesh2d(){
 
 	PetscErrorCode ierr;
 	PetscScalar             **qq,**qq_cont,**qq_rho,**TT,**qq_H,**qq_strain;
@@ -459,6 +460,822 @@ PetscErrorCode Swarm2Mesh(){
 	}
 	
 	
+	PetscFunctionReturn(0);
+	
+}
+
+PetscErrorCode Swarm2Mesh3d(){
+
+	PetscErrorCode ierr;
+	PetscScalar             ***qq,***qq_cont,***qq_rho,***TT,***qq_H,***qq_strain;
+	PetscScalar				***qq_strain_rate;
+	
+	ierr = VecSet(geoq,0.0);CHKERRQ(ierr);
+	ierr = VecSet(geoq_rho,0.0);CHKERRQ(ierr);
+	ierr = VecSet(geoq_cont,0.0);CHKERRQ(ierr);
+	ierr = VecSet(geoq_H,0.0);CHKERRQ(ierr);
+	ierr = VecSet(geoq_strain,0.0);CHKERRQ(ierr);
+	ierr = VecSet(geoq_strain_rate,0.0);CHKERRQ(ierr);
+	
+	ierr = VecZeroEntries(local_geoq);CHKERRQ(ierr);
+	ierr = VecZeroEntries(local_geoq_rho);CHKERRQ(ierr);
+	ierr = VecZeroEntries(local_geoq_H);CHKERRQ(ierr);
+	ierr = VecZeroEntries(local_geoq_strain);CHKERRQ(ierr);
+	ierr = VecZeroEntries(local_geoq_strain_rate);CHKERRQ(ierr);
+	
+	ierr = DMGlobalToLocalBegin(da_Thermal,geoq,INSERT_VALUES,local_geoq);
+	ierr = DMGlobalToLocalEnd(  da_Thermal,geoq,INSERT_VALUES,local_geoq);
+	
+	ierr = DMGlobalToLocalBegin(da_Thermal,geoq_rho,INSERT_VALUES,local_geoq_rho);
+	ierr = DMGlobalToLocalEnd(  da_Thermal,geoq_rho,INSERT_VALUES,local_geoq_rho);
+	
+	ierr = DMGlobalToLocalBegin(da_Thermal,geoq_H,INSERT_VALUES,local_geoq_H);
+	ierr = DMGlobalToLocalEnd(  da_Thermal,geoq_H,INSERT_VALUES,local_geoq_H);
+	
+	ierr = DMGlobalToLocalBegin(da_Thermal,geoq_strain,INSERT_VALUES,local_geoq_strain);
+	ierr = DMGlobalToLocalEnd(  da_Thermal,geoq_strain,INSERT_VALUES,local_geoq_strain);
+
[…]
