### 2021-11-29 71915e7
addition of flag initial_strain_gaussian_width
