+++ src/DM1_v.cpp
@@ -4,8 +4,10 @@
 #include <petsctime.h>
 
 PetscErrorCode DMDAGetElementCorners(DM da,PetscInt *sx,PetscInt *sz,PetscInt *mx,PetscInt *mz);
+PetscErrorCode DMDAGetElementCorners3d(DM da,PetscInt *sx,PetscInt *sy, PetscInt *sz,PetscInt *mx, PetscInt *my, PetscInt *mz);
 
 PetscReal montaKeVeloc_simplif_2d(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
+PetscReal montaKeVeloc_simplif_3d(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
 
 PetscErrorCode write_veloc_3d(int cont, PetscInt binary_out);
 PetscErrorCode write_veloc_cond(int cont, PetscInt binary_out);
@@ -47,6 +49,7 @@ extern PetscReal *Ke_veloc_final;
 extern PetscReal *Ke_veloc_general;
 
 extern double dx_const;
+extern double dy_const;
 extern double dz_const;
 
 extern double depth;
@@ -122,7 +125,7 @@ extern PetscReal theta_FSSA;
 
 extern PetscInt periodic_boundary;
 
-PetscErrorCode AssembleA_Veloc(Mat A,Mat AG,DM veloc_da, DM temper_da){
+PetscErrorCode AssembleA_Veloc2d(Mat A,Mat AG,DM veloc_da, DM temper_da){
 
 	PetscErrorCode         ierr;
 
@@ -373,7 +376,271 @@ PetscErrorCode AssembleA_Veloc(Mat A,Mat AG,DM veloc_da, DM temper_da){
 	PetscFunctionReturn(0);
 }
 
-PetscErrorCode AssembleF_Veloc(Vec F,DM veloc_da,DM drho_da,Vec FP){
+PetscErrorCode AssembleA_Veloc3d(Mat A,Mat AG,DM veloc_da, DM temper_da){
+
+	PetscErrorCode         ierr;
+	
+	PetscInt i,j,k;
+	
+	PetscInt               M,N,P;
+	
+	MatStencil indr[T_NE],ind1[1],ind[V_GT], indp[1];
+	
+	PetscScalar u[V_GT*V_GT],val_cond[1];
+	
+	Stokes3d					***pc;
+	
+	ierr = VecZeroEntries(local_Precon);CHKERRQ(ierr);
+	ierr = DMDAVecGetArray(veloc_da,local_Precon,&pc);CHKERRQ(ierr);
+	
+	Stokes3d					***VVC;
+	
+	ierr = VecZeroEntries(local_VC);CHKERRQ(ierr);
+	
+	ierr = DMGlobalToLocalBegin(veloc_da,Veloc_Cond,INSERT_VALUES,local_VC);
+	ierr = DMGlobalToLocalEnd(  veloc_da,Veloc_Cond,INSERT_VALUES,local_VC);
+	
+	ierr = DMDAVecGetArray(veloc_da,local_VC,&VVC);CHKERRQ(ierr);
+	
+	PetscInt       sx,sy,sz,mmx,mmy,mmz;
+	
+	ierr = DMDAGetCorners(veloc_da,&sx,&sy,&sz,&mmx,&mmy,&mmz);CHKERRQ(ierr);
+	
[…]

[…]
