### 2022-01-07 #3551 @jedbrown
@aguspesce Yikes, I'm terribly sorry about my silence. Thanks for your effort here; it looks good. I see `CONTRIBUTING.md` says to run `make test`, but it actually needs `make test_mandyoc` (or change the Makefile target). The scripts also look inconsistent in where they get `mpirun` (PETSc directory or `$MPI_PATH`). It's probably best to make this a run-time option, but you have some choices
* `make getmpiexec` with your Makefile
* `$(MPIEXEC)` inside your Makefile
* `$(dirname $(pkg-config --variable=ccompiler $PETSC_DIR/$PETSC_ARCH/lib/pkgconfig/petsc.pc))/mpiexec` (less reliable)

Here's a patch to make the test run out of the box, but the scripts still use the fragile style.
```diff
diff --git i/Makefile w/Makefile
index e0e91de..3d2f3e6 100644
--- i/Makefile
+++ w/Makefile
@@ -26,8 +26,6 @@ SOURCEC = $(SRC)/main.cpp \
 	$(SRC)/sp.cpp
 OBJECTS = $(SOURCEC:%.cpp=%.o)

-MPI_PATH = ${PETSC_DIR}/${PETSC_ARCH}/bin
-
 help:
 	@echo ""
 	@echo "Commands:"
@@ -40,7 +38,7 @@ help:
 test_mandyoc:

 	@echo "Run MANDYOC test may take several minutes.."
-	cd test/testing_data/ ; ${MPI_PATH}/mpirun -n 2 ../../mandyoc
+	cd test/testing_data/ ; ${MPIEXEC} -n 2 ../../mandyoc
 	pytest -v test/testing_result.py

 # Build Mandyoc
```
I ran this and get one test failure (`0-velocity`):
```

field = 'velocity', step = 0

    @pytest.mark.parametrize("field", fields)
    @pytest.mark.parametrize("step", steps)
    def test_result(field, step):
        """ """
        filename = f"{field}_{step}" + ".txt"
        output = read(test_path / filename)
        expected = read(expected_path / filename)

>       npt.assert_allclose(output, expected, rtol=1e-5)
E       AssertionError:
E       Not equal to tolerance rtol=1e-05, atol=0
E
E       Mismatched elements: 1 / 94978 (0.00105%)
E       Max absolute difference: 1.e-16
E       Max relative difference: 1.10695082e-05
E        x: array([0., 0., 0., ..., 0., 0., 0.])
E        y: array([0., 0., 0., ..., 0., 0., 0.])

test/testing_result.py:57: AssertionError
```

I'll find reviewers.
