### 2022-01-07 #3551 @jedbrown
@gassmoeller @rbeucher :wave: Would you be available to review this submission to JOSS?
