### 2022-01-11 0f859b9
Fix an error in the test section in the CONTRIBUTING file.
Change `make test` for `make test_mandyoc`
