### 2022-01-11 2a94f7f
Change variables in the Makefile
Eliminate MPI_PATH and create MPIEXEC for the test
