### 2022-01-11 fee18e8
Improve Makefile changing the MPI variable for the test (#39)
* Fix an error in the test section in the CONTRIBUTING file.

Change `make test` for `make test_mandyoc`

* Change variables in the Makefile

Eliminate MPI_PATH and create MPIEXEC for the test

* Eliminate a space
