### 2022-01-12 #3551 @victorsacek
Hello @jedbrown

Thank you very much for your attention.

We modified the Makefile, changing the MPI variables for the test, as you suggested.

Additionally, in the Statement of need, we added one sentence explaining that one advantage of Mandyoc is the possibility to incorporate variable boundary conditions in space and time, appropriate to simulate different pulses of tectonism in one model run.

About the failure test, we tried different machines but we couldn't reproduce this failure. Please, if possible, could you send us more details about the problem?

Best regards
