### 2022-01-15 #3551 @jedbrown
@victorsacek Thanks for your updates. I think you'll find that other software supports BCs that depend on space and time. I think this section should be specific in placing present work in the context of related work. In this case, I think it involves citing some alternatives and explaining the relative merits.

Also, in the `Makefile`, it's better not to change `MPIEXEC`, which has already been defined in the makefile you included. Your redefinition prevents use of system MPI (i.e., when not using `--download-mpich` or `--download-openmpi`).

Thanks for being willing to review, @rbeucher! I'll add you and start the review. I have some emails out and will find a second reviewer as well. Of course I'd be thrilled if @bangerth or @tjhei are available. And I totally understand this present state of being burried @gassmoeller; thanks for your reply.
