### 2022-01-15 #4070 @whedon
[REVIEW]: Mandyoc: A finite element code to simulate thermochemical convection in parallel
**Submitting author:** @victorsacek (<a href="http://orcid.org/0000-0001-9598-5081">Victor Sacek</a>)
**Repository:** https://github.com/ggciag/mandyoc
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.6

- [x] #4128
