### 2022-01-18 ggciag/mandyoc#40 @rafaelmds
Missing DOIs
Review citation DOIs for JOSS paper.

[https://github.com/openjournals/joss-reviews/issues/3551](https://github.com/openjournals/joss-reviews/issues/3551)
