### 2022-01-18 ggciag/mandyoc#41 @rafaelmds
Revert and fix MPIEXEC in Makefile
As [suggested](https://github.com/openjournals/joss-reviews/issues/3551#issuecomment-1013612980) by @jedbrown, its better to not change MPIEXEC in Makefile (otherwise the system might not be able to use system MPI.

[https://github.com/openjournals/joss-reviews/issues/3551](https://github.com/openjournals/joss-reviews/issues/3551)
