### 2022-01-18 1c3ab16
add dois and url to references
