+++ paper/paper.bib
@@ -1,12 +1,13 @@
 @article{crameri2012comparison,
-  title={A comparison of numerical surface topography calculations in geodynamic modelling: an evaluation of the ‘sticky air’method},
+  title={A comparison of numerical surface topography calculations in geodynamic modelling: an evaluation of the ‘sticky air’ method},
   author={Crameri, F and Schmeling, H and Golabek, GJ and Duretz, T and Orendt, R and Buiter, SJH and May, DA and Kaus, BJP and Gerya, TV and Tackley, PJ},
   journal={Geophysical Journal International},
   volume={189},
   number={1},
   pages={38--54},
   year={2012},
-  publisher={Blackwell Publishing Ltd Oxford, UK}
+  publisher={Blackwell Publishing Ltd Oxford, UK},
+  doi={10.1111/j.1365-246X.2012.05388.x}
 }
 
 @article{kaus2010stabilization,
@@ -17,7 +18,8 @@
   number={1-2},
   pages={12--20},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.pepi.2010.04.007}
 }
 
 @inproceedings{may2017dmswarm,
@@ -25,7 +27,8 @@
   author={May, Dave A and Knepley, Matthew G},
   booktitle={EGU General Assembly Conference Abstracts},
   pages={10133},
-  year={2017}
+  year={2017},
+  url={https://ui.adsabs.harvard.edu/abs/2017EGUGA..1910133M/abstract}
 }
 
 @Misc{petsc-web-page,
@@ -70,7 +73,8 @@
   volume={459},
   pages={48--57},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.epsl.2016.11.026}
 }
 
 @article{salazar2021lateral,
@@ -80,7 +84,8 @@
   volume={145},
   pages={101830},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jog.2021.101830}
 }
 
 @article{van1997comparison,
@@ -91,5 +96,6 @@
   number={B10},
   pages={22477--22495},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/97JB01353}
 }
