+++ paper/paper.bib
@@ -63,7 +63,8 @@
   Editor = "E. Arge and A. M. Bruaset and H. P. Langtangen",
   Pages = "163--202",
   Publisher = "Birkh{\"{a}}user Press",
-  Year = "1997"
+  Year = "1997",
+  doi={10.1007/978-1-4612-1986-6_8}
 }
 
 @article{sacek2017post,
