### 2022-01-18 480ae4b
Fix DOIs (#43)
* remove not used references

* add missing doi and url to references
