### 2022-01-18 330fa21
fix description of second invariant term
Co-authored-by: victorsacek <sacek@usp.br>
