### 2022-01-18 4976fdf
docs: fix expression for viscosity component (#37)
* fix expression for viscosity component in docs

* fix description of second invariant term

Co-authored-by: victorsacek <sacek@usp.br>

Co-authored-by: victorsacek <sacek@usp.br>
