### 2022-01-18 6b83497
add note about MPIEXEC variable in mandyoc-installation doc
