### 2022-01-18 0914456
Add notes about mpiexec variable (#45)
* add notes about MPIEXEC variable in mandyoc-installation doc and contributing section
