### 2022-02-03 #4070 @rbeucher
@victorsacek

Thanks for your submission. I have installed the code and I am currently running the examples. So far so good.
I have opened a series of issues on the repository issue tracker with some comments and suggestions on how to make the submission better.

The code compiles and run as described.
It needs some clearer examples that cover the range of functionalities describe in the documentation/paper.
The documentation is incomplete. It also needs simple tutorials to get the user started.
I would strongly suggest some basic automated tests....

The paper needs some work both in terms of structure and content. I think the content needs to be aligned with the
title. See my comments on the issue tracker.
