### 2022-02-07 4f135e5
Improve the build and installation of mandyoc.
