### 2022-02-07 a7f7e82
Delete the default path to petsc as ~/petsc in the Makefile
