### 2022-02-07 1a60976
Improve the installation instruction according to the updated Makefile - Remove the default path for pests a t ~ - Add instruction to  add a symlinks of mpirun - Add instruction to create an env variable for PETSC_DIR before build Mandyoc
