### 2022-02-08 8aa3087
Changed incorrect image; changed param.txt file in the documentation; changed van Keken's Python script to create proper param.txt file
