+++ docs/files/src/param.txt
@@ -1,82 +1,111 @@
+
 # Geometry
-nx                        = 281        # N. of elements in the longitudinal direction
-nz                        = 85         # N. of elements in the vertical direction
-lx                        = 2800000.0  # Extent in the longitudinal direction
-lz                        = 850000.0   # Extent in the vertical direction
+nx                                  = 81            # n. of nodes in the longitudinal direction
+nz                                  = 81            # n. of nodes in the vertical direction
+lx                                  = 91420.0       # extent in the longitudinal direction
+lz                                  = 100000.       # extent in the vertical direction
+particles_per_element_x             = 0             # default is 0
+particles_per_element_z             = 0             # default is 0
 
 # Simulation options
-# solver                  = direct     # [direct/iterative]
-use_multigrid             = 1          # Enable multigrid
-# residual_norm_iter      = 1.0e-5
-# residual_norm_uzawa     = 1.0e-4
+multigrid                           = 1             # ok -> soon to be on the command line only
+solver                              = direct        # default is direct [direct/iterative]
+denok                               = 1.0e-15       # default is 1.0e-4
+particles_per_element               = 1100          # default is 81
+particles_perturb_factor            = 0.0           # default is 0.5 [values are between 0 and 1]
+rtol                                = 1.0e-5        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
+RK4                                 = Euler         # default is Euler [Euler/Runge-Kutta]
+Xi_min                              = 1.0e-14       # default is 1.0e-14
+random_initial_strain               = 0.0           # default is 0.0
+pressure_const                      = -1.0          # default is -1.0 (not used)
+initial_dynamic_range               = False         # default is False [True/False]
+periodic_boundary                   = False         # default is False [True/False]
+high_kappa_in_asthenosphere         = False         # default is False [True/False]
+K_fluvial                           = 2.0e-7        # default is 2.0e-7
+m_fluvial                           = 1.0           # default is 1.0
+sea_level                           = 0.0           # default is 0.0
+basal_heat                          = -1.0          # default is -1.0
 
-# Time constrains for the simulation
-step_maxkjdfd             = 7000       # Maximum time-step of the simulation
-step_print                = 10         # Make file every <step_print>
-time_max                  = 14e9       # Maximum time of the simulation [s]
-dt_max                    = 10.0e3     # Maximum time between steps of the simulation [s]
+# Surface processes
+sp_surface_tracking                 = False         # default is False [True/False]
+sp_surface_processes                = False         # default is False [True/False]
+sp_dt                               = 0             # default is 0
+sp_d_c                              = 0             # default is 0
+plot_sediment                       = False         # default is False [True/False]
+a2l                                 = True          # default is True [True/False]
 
-# Values for the viscosity
-visc_ref                  = 1.0e+21    # Reference viscosity [Pa.s]
-visc_max                  = 1.0e+23    # Maximum viscosity [Pa.s]
-visc_min                  = 1.0e+19    # Minimum viscosity [Pa.s]
+free_surface_stab                   = True          # default is True [True/False]
+theta_FSSA                          = 0.5           # default is 0.5 (only relevant when free_surface_stab = True)
 
-# External ASCII inputs
-interfaces_from_ascii     = True       # Use 'interfaces.txt' ASCII file
-# temperature_from_ascii  = True       # Use 'input_temperature_0.txt' ASCII file
-velocity_from_ascii       = False      # Use 'input_velocity_0.txt' ASCII file
+# Time constrains
+step_max                            = 4000          # maximum time-step of the simulation
[…]
