### 2022-02-10 #4070 @psanan
@victorsacek, thanks for the submission! @rbeucher has opened several issues on the Mandyoc tracker, which cover the majority of points I wanted to make, at least generally, so I will just add my own comments to those issues where appropriate, and open new ones (limited in scope) if one of those issues is resolved but I still think some further change is required for acceptance.
