### 2022-02-10 a02f4e1
Fix typo in Makefile
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
