### 2022-02-10 fe266a6
Fix typo in the help section
Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
