### 2022-02-14 90ec5b0
Create a jupyter notebook to generate and plot the continental rift example
Create the notebook to replace the python scripts.
