+++ examples/continental_rift/generate_input_files.py
@@ -1,414 +1,0 @@
-"""
-This example simulates the evolution of divergent margins, taking into account the plastic rheology and the sin-rift geodynamics
-"""
-
-import numpy as np
-import matplotlib.pyplot as plt
-
-# total model horizontal extent (m)
-Lx = 1600 * 1.0e3
-# total model vertical extent (m)
-Lz = 300 * 1.0e3
-# number of points in horizontal direction
-Nx = 1601
-# number of points in vertical direction
-Nz = 301
-# thickness of sticky air layer (m)
-H_sa = 40 * 1.0e3
-# thickness of lower crust (m)
-H_lower_crust = 20 * 1.0e3
-# thickness of upper crust (m)
-H_upper_crust = 20 * 1.0e3
-# total thickness of lithosphere (m)
-H_litho = 130 * 1.0e3
-# seed depth bellow base of lower crust (m)
-seed_depth = 13 * 1.0e3
-
-x = np.linspace(0, Lx, Nx)
-z = np.linspace(Lz, 0, Nz)
-X, Z = np.meshgrid(x, z)
-
-
-##############################################################################
-# Interfaces (bottom first)
-##############################################################################
-interfaces = {
-    "litho": np.ones(Nx) * (H_litho + H_sa),
-    "seed_base": np.ones(Nx) * (seed_depth + H_lower_crust + H_upper_crust + H_sa),
-    "seed_top": np.ones(Nx) * (seed_depth + H_lower_crust + H_upper_crust + H_sa),
-    "lower_crust": np.ones(Nx) * (H_lower_crust + H_upper_crust + H_sa),
-    "upper_crust": np.ones(Nx) * (H_upper_crust + H_sa),
-    "air": np.ones(Nx) * (H_sa),
-}
-
-# seed thickness (m)
-H_seed = 6 * 1.0e3
-# seed horizontal position (m)
-x_seed = 750 * 1.0e3
-# seed: number of points of horizontal extent
-n_seed = 2
-
-interfaces["seed_base"][
-    int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
-] = (
-    interfaces["seed_base"][
-        int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
-    ]
-    + H_seed // 2
-)
-interfaces["seed_top"][
-    int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
-] = (
-    interfaces["seed_top"][
-        int(Nx * x_seed // Lx - n_seed // 2) : int(Nx * x_seed // Lx + n_seed // 2)
-    ]
-    - H_seed // 2
-)
-
-Huc = 2.5e-6 / 2700.0
-Hlc = 0.8e-6 / 2800.0
-
-# Create the interface file
-with open("interfaces.txt", "w") as f:
-    layer_properties = f"""
-        C   1.0       1.0        0.1        1.0        1.0         1.0         1.0
-        rho 3378.0    3354.0     3354.0     3354.0     2800.0      2700.0      1.0
-        H   0.0       9.0e-12    9.0e-12    9.0e-12    {Hlc}       {Huc}       0.0
-        A   1.393e-14 2.4168e-15 2.4168e-15 2.4168e-15 8.574e-28   8.574e-28   1.0e-18
-        n   3.0       3.5        3.5        3.5        4.0         4.0         1.0
-        Q   429.0e3   540.0e3    540.0e3    540.0e3    222.0e3     222.0e3     0.0
-        V   15.0e-6   25.0e-6    25.0e-6    25.0e-6    0.0         0.0         0.0
-    """
-
-    for line in layer_properties.split("\n"):
[…]
