### 2022-02-14 8799663
Remove readme file. This text is in the jupyter notebook now.
