### 2022-02-15 7c8c099
Printing particles output just after the first migration
