### 2022-02-15 d4ff508
Add error message if PETSC_DIR is not defined and improve variable names for directories
