### 2022-02-16 b2d9bc8
Remove BINDIR as a requirement in the installation
