+++ docs/files/src/param.txt
@@ -1,6 +1,6 @@
 
 # Geometry
-nx                                  = 81            # n. of nodes in the longitudinal direction
+nx                                  = 81            # n. of nodes in the horizontal direction
 nz                                  = 81            # n. of nodes in the vertical direction
 lx                                  = 91420.0       # extent in the longitudinal direction
 lz                                  = 100000.       # extent in the vertical direction
