### 2022-02-16 ef1942a
Removed unused line
Co-authored-by: victorsacek <sacek@usp.br>
