### 2022-02-16 bcd08b5
Changed Python code to print chenged word
Co-authored-by: victorsacek <sacek@usp.br>
