### 2022-02-16 9d87148
Removed line and changed word
Co-authored-by: victorsacek <sacek@usp.br>
