### 2022-02-16 a0bb93e
Removed line and changed word in the param.txt file
Co-authored-by: victorsacek <sacek@usp.br>
