### 2022-02-16 27c2ff4
Inprove the documentation about how build and install Mandyoc
