### 2022-02-16 c389814
Add docker images and python requirements file needed to build images (to support running examples and tests)
