### 2022-02-18 9a2f6e5
geoq_on removed from param.txt; added default value to geoq_on in the header.h file; improved description of the parameters inside the param.txt file
