+++ docs/files/src/param.txt
@@ -1,110 +1,101 @@
-
 # Geometry
 nx                                  = 81            # n. of nodes in the horizontal direction
 nz                                  = 81            # n. of nodes in the vertical direction
 lx                                  = 91420.0       # extent in the horizontal direction
 lz                                  = 100000.       # extent in the vertical direction
-particles_per_element_x             = 0             # default is 0
-particles_per_element_z             = 0             # default is 0
 
 # Simulation options
-solver                              = direct        # default is direct [direct/iterative]
-denok                               = 1.0e-15       # default is 1.0e-4
-particles_per_element               = 1100          # default is 81
-particles_perturb_factor            = 0.0           # default is 0.5 [values are between 0 and 1]
-rtol                                = 1.0e-5        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
-RK4                                 = Euler         # default is Euler [Euler/Runge-Kutta]
-Xi_min                              = 1.0e-14       # default is 1.0e-14
-random_initial_strain               = 0.0           # default is 0.0
-pressure_const                      = -1.0          # default is -1.0 (not used)
-initial_dynamic_range               = False         # default is False [True/False]
-periodic_boundary                   = False         # default is False [True/False]
-high_kappa_in_asthenosphere         = False         # default is False [True/False]
-K_fluvial                           = 2.0e-7        # default is 2.0e-7
-m_fluvial                           = 1.0           # default is 1.0
-sea_level                           = 0.0           # default is 0.0
+solver                              = direct        # set the solver to be direct or iterative, default is direct [direct/iterative]
+denok                               = 1.0e-15       # tolerance criterion for the Uzawa's scheme, default is 1.0e-4
+rtol                                = 1.0e-5        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0e-5
+RK4                                 = Euler         # particles advection method, default is Euler [Euler/Runge-Kutta]
+Xi_min                              = 1.0e-14       # tolerance criterion for the convergence of the non-linear flow, default is 1.0e-14
+random_initial_strain               = 0.0           # non-dimensional value for the initial strain perturbation for the entire domain, default is 0.0 
+pressure_const                      = -1.0          # set constant pressure value for the domain (relevant only when 2D is plain view), default is -1.0 (i.e. not used) [Pa]
+initial_dynamic_range               = False         # method to smoothen convergence of the velocity field in scenarios with wide viscosity range, default is False [True/False]
+periodic_boundary                   = False         # allows simulation with periodic boundary in the horizontal direction, default is False [True/False]
+high_kappa_in_asthenosphere         = False         # mimics high heat transport in the asthenosphere increasing its thermal diffusivity coefficient, default is False [True/False]
+
+
 basal_heat                          = -1.0          # default is -1.0
 
-# Surface processes
-sp_surface_tracking                 = False         # default is False [True/False]
-sp_surface_processes                = False         # default is False [True/False]
-sp_dt                               = 0             # default is 0
-sp_d_c                              = 0             # default is 0
-plot_sediment                       = False         # default is False [True/False]
-a2l                                 = True          # default is True [True/False]
+# Particles options
+particles_per_element               = 1100          # default is 81
+particles_per_element_x             = 0             # n. of Lagrangian particles in the horizontal direction, default is 0 for automatic calculation
+particles_per_element_z             = 0             # n. of Lagrangian particles in the vertical direction, default is 0 for automatic calculation
+particles_perturb_factor            = 0.0           # default is 0.5 [values are between 0 and 1]
 
-free_surface_stab                   = True          # default is True [True/False]
[…]
