### 2022-02-22 38f5ac7
improved param.txt example file, added description to parameters in the Parameter file section, added input examples and explanation (multi_veloc.txt and scale_bcv.txt)
