+++ docs/files/input.rst
@@ -107,8 +107,6 @@ ASCII velocities files
 ..
    For both 2-D and 3-D grids, the ``input_velocity.txt`` file possesses four lines of headers followed by the velocity data for each node of the model.
 
-For a 2-D grid, the ``input_velocity_0.txt`` file possesses four lines of headers followed by the velocity data for each node of the model.
-
 An initial velocity file allows an initial velocity field to be established so the simulated fluid presents an initial momentum. Respecting the conservation of mass equation when designing the initial velocity field is fundamental so the simulation can run properly.
 
 Additionally, the velocity boundary conditions defined in the ``param.txt`` (see the :doc:`parameter file section<parameter-file>`) will be used to define the normal and tangential velocities on the faces of the model during the simulation.
@@ -116,11 +114,35 @@ Additionally, the velocity boundary conditions defined in the ``param.txt`` (see
 2-D initial velocity
 ********************
 
-For a 2-D grid, the velocity must be defined for both :math:`x` and :math:`y` directions. Taking the example in the file below, the ``input_velocity_0.txt`` file must be written in such a way that the values for the velocity in the :math:`x` direction :math:`v_x` and the velocity in the :math:`y` direction :math:`v_y` for the node :math:`(x_0,y_0)` corresponds to lines 5 and 6, respectively, the velocity values for the :math:`(x_1,y_0)` corresponds to lines 7 and 8, and so on until :math:`(x_{nx-1},y_0)`. Once all the nodes in the :math:`y_0` are written, the values for all :math:`x_n` are inserted for :math:`y_1`, and so on until :math:`y_{ny-1}`.
+When ``velocity_from_ascii = True`` is set in the ``param.txt`` file, an initial velocity field must be provided in an ASCII file called ``input_velocity_0.txt``.
+
+For a 2-D grid, the ``input_velocity_0.txt`` file possesses four lines of headers followed by the velocity data for each node of the model. The velocity must be defined for both :math:`x` and :math:`y` directions. Taking the example in the file below, the ``input_velocity_0.txt`` file must be written in such a way that the values for the velocity in the :math:`x` direction :math:`v_x` and the velocity in the :math:`y` direction :math:`v_y` for the node :math:`(x_0,y_0)` corresponds to lines 5 and 6, respectively, the velocity values for the :math:`(x_1,y_0)` corresponds to lines 7 and 8, and so on until :math:`(x_{nx-1},y_0)`. Once all the nodes in the :math:`y_0` are written, the values for all :math:`x_n` are inserted for :math:`y_1`, and so on until :math:`y_{ny-1}`.
 
 .. literalinclude:: src/initial-velocity-2d.txt
    :language: text
    :linenos:
+
+2-D multiple initial velocities
+*******************************
+
+The user can provide multiple velocity fields to be set in different instants of the simulation. By setting ``multi_velocity = True`` in the ``param.txt`` file, an ASCII file called ``multi_veloc.txt`` must be provided following the structure shown in the example file below. The first line of the file contains the amount of instants where a different velocity field will be used and the other lines contain the time, in millions of years, that each velocity field will be adopted. In the example below, a different velocity field will be used at three instants, the files must follow the structure presented in the last section and should be named in sequence: ``input_velocity_1.txt``, ``input_velocity_2.txt`` and ``input_velocity_3.txt``. The names of the files are labeled after the three different instants set in the first line of the ``multi_veloc.txt`` file.
+
+.. literalinclude:: src/multi_veloc.txt
+   :language: text
+   :linenos:
+
+.. note::
[…]
