+++ docs/files/src/param.txt
@@ -12,15 +12,13 @@ RK4                                 = Euler         # particles advection method
 Xi_min                              = 1.0e-14       # tolerance criterion for the convergence of the non-linear flow, default is 1.0e-14
 random_initial_strain               = 0.0           # non-dimensional value for the initial strain perturbation for the entire domain, default is 0.0 
 pressure_const                      = -1.0          # set constant pressure value for the domain (relevant only when 2D is plain view), default is -1.0 (i.e. not used) [Pa]
-initial_dynamic_range               = False         # method to smoothen convergence of the velocity field in scenarios with wide viscosity range, default is False [True/False]
+initial_dynamic_range               = False         # method to smoothen convergence of the velocity field in scenarios with wide viscosity range, default is False [True/False], Gerya (2019)
 periodic_boundary                   = False         # allows simulation with periodic boundary in the horizontal direction, default is False [True/False]
 high_kappa_in_asthenosphere         = False         # mimics high heat transport in the asthenosphere increasing its thermal diffusivity coefficient, default is False [True/False]
-
-
-basal_heat                          = -1.0          # default is -1.0
+basal_heat                          = -1.0          # set basal heat flux value, default is -1.0 (i.e. not used) [W/m^2]
 
 # Particles options
-particles_per_element               = 1100          # default is 81
+particles_per_element               = 1100          # n. of Lagrangian particles in each element, default is 81
 particles_per_element_x             = 0             # n. of Lagrangian particles in the horizontal direction, default is 0 for automatic calculation
 particles_per_element_z             = 0             # n. of Lagrangian particles in the vertical direction, default is 0 for automatic calculation
 particles_perturb_factor            = 0.0           # default is 0.5 [values are between 0 and 1]
@@ -32,8 +30,8 @@ sp_surface_processes                = False         # allows erosion and sedimen
 sp_dt                               = 0             # time step for surface processes simulation, default is 0 [years]
 a2l                                 = True          # allows the conversion of air particles to land particles during sedimentation, default is True [True/False]
 sp_mode                             = 1             # specify the surface processes method, default is 1
-free_surface_stab                   = True          # set if the free surface stabilization algorithm (FSSA) is used, default is True [True/False]
-theta_FSSA                          = 0.5           # weight of the influence of the FSSA method, default is 0.5 (only relevant when free_surface_stab = True)
+free_surface_stab                   = True          # set if the free surface stabilization algorithm (FSSA) is used, default is True [True/False], Kaus et al. (2010)
+theta_FSSA                          = 0.5           # weight of the influence of the FSSA method, default is 0.5 (only relevant when <free_surface_stab> is True)
 sticky_blanket_air                  = False         # allows the increase of viscosity for the first air layer of particles, default is False [True/False]
 precipitation_profile_from_ascii    = False         # set if precipitation profile along the horizontal axis is read from an ASCII file, default is False [True/False]
 climate_change_from_ascii           = False         # if True, re-scales through time the precipitation profile using an ASCII file, default is False [True/False]
@@ -56,26 +54,26 @@ viscosity_mean_method               = harmonic      # defines method do calculat
 viscosity_dependence                = pressure      # defines if viscosity depends on pressure or depth, default is depth [pressure/depth]
 
 # External ASCII inputs/outputs
-interfaces_from_ascii               = True          # set if interfaces between lithologies are read from an ASCII file, default is False [True/False]
-n_interfaces                        = 1             # set the number of interfaces to be read from the interfaces ASCII file
-variable_bcv                        = False         # set if velocity boundary condition changes according to an ASCII file, default is False [True/False]
-temperature_from_ascii              = False         # set if initial temperature is read from an ASCII file, default is False [True/False]
[…]
