+++ docs/files/refs.bib
@@ -155,6 +155,24 @@
   publisher={Elsevier Amsterdam}
 }
 
+@book{gerya2019,
+  title={Introduction to numerical geodynamic modelling},
+  author={Gerya, Taras},
+  year={2019},
+  publisher={Cambridge University Press}
+}
+
+@article{kaus2010,
+  title={A stabilization algorithm for geodynamic numerical simulations with a free surface},
+  author={Kaus, Boris JP and M{\"u}hlhaus, Hans and May, Dave A},
+  journal={Physics of the Earth and Planetary Interiors},
+  volume={181},
+  number={1-2},
+  pages={12--20},
+  year={2010},
+  publisher={Elsevier}
+}
+
 @article{drucker-prager1952,
   title={Soil mechanics and plastic analysis or limit design},
   author={Drucker, Daniel Charles and Prager, William},

[…]
