### 2022-02-23 5c7709c
Improve installation instruction in the docimentation
Add a note about the ~/.local/bin folder

Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>
