### 2022-02-25 3b1cd11
Change the variable name for the installation
