### 2022-02-25 5786f6a
Add a note in the installation documentation about how to set the installation path
