### 2022-03-02 0f574f4
Remove default petsc path in the Makefile and istallation (#55)
* Improve the build and installation of mandyoc.

* Delete the default path to petsc as ~/petsc in the Makefile

* Change test_mandyoc for test in Contributing file

* Improve the installation instruction according to the updated Makefile
- Remove the default path for pests a t ~
- Add instruction to  add a symlinks of mpirun
- Add instruction to create an env variable for PETSC_DIR before build Mandyoc

* Fix typo in Makefile

Co-authored-by: Rafael Silva <rafaelmds@users.noreply.github.com>

* Fix typo in the help section

Co-au
[…]
