+++ paper/paper.bib
@@ -89,6 +89,16 @@
   doi={10.1016/j.jog.2021.101830}
 }
 
+@article{silva2022influence,
+  title={Influence of surface processes on post-rift faulting during divergent margins evolution},
+  author={Silva, Rafael Monteiro da and Sacek, V},
+  journal={Tectonics},
+  volume={41},
+  pages={e2021TC006808},
+  year={2022},
+  publisher={Wiley Online Library}
+}
+
 @article{van1997comparison,
   title={A comparison of methods for the modeling of thermochemical convection},
   author={Van Keken, PE and King, SD and Schmeling, H and Christensen, UR and Neumeister, D and Doin, M-P},
@@ -99,4 +109,62 @@
   year={1997},
   publisher={Wiley Online Library},
   doi={10.1029/97JB01353}
+}
+
+
+
+
+
+@Article{wolfgang_bangerth_2021_5131909, 
+	author = "Bangerth, W. and Dannberg, J. and Fraters, M. and Gassmoeller, R. and Glerum, A. and Heister, T. and Naliboff, J.", 
+	title="ASPECT v2.3.0", 
+	year="2021", 
+	publisher="Zenodo", 
+	optkeywords="ASPECT", 
+	doi="http://doi.org/10.5281/zenodo.5131909", 
+	opturl="https://doi.org/10.5281/zenodo.5131909"}
+
+@Article{Bangerth2021, 
+	author = "Bangerth, W. and Dannberg, J. and Fraters, M. and Gassmoeller, R. and Glerum, A. and Heister, T. and Naliboff, J.", 
+	title="ASPECT: Advanced Solver for Problems in Earth{\textquoteright}s ConvecTion, User Manual", 
+	year="2021", 
+	optkeywords="ASPET", 
+	doi="http://doi.org/10.6084/m9.figshare.4865333.v8", 
+	opturl="https://figshare.com/articles/journal_contribution/ASPECT_Advanced_Solver_for_Problems_in_Earth_s_ConvecTion_User_Manual/4865333"}
+
+@Article{KHB12, 
+	author = "Kronbichler, M. and Heister, T. and Bangerth, W.", 
+	title="High accuracy mantle convection simulation through modern numerical methods", 
+	year="2012", 
+	journal="Geophysical Journal International", 
+	volume="191", 
+	number="1", 
+	pages="12-29", 
+	optkeywords="ASPECT", 
+	issn="0956540X", 
+	doi="http://doi.org/10.1111/j.1365-246X.2012.05609.x", 
+	opturl="http://gji.oxfordjournals.org/cgi/doi/10.1111/j.1365-246X.2012.05609.x"}
+
+@Article{aspect2017, 
+	author = "Heister, T. and Dannberg, J. and Gassm{\"o}ller, R. and Bangerth, W.", 
+	title="High accuracy mantle convection simulation through modern numerical methods - II: realistic models and problems", 
+	year="2017", 
+	journal="Geophysical Journal International", 
+	volume="210", 
+	number="2", 
+	pages="833-851", 
+	optkeywords="ASPECT", 
+	issn="0956-540X", 
+	doi="http://doi.org/10.1093/gji/ggx195", 
+	opturl="https://academic.oup.com/gji/article-lookup/doi/10.1093/gji/ggx195"}
+
+
+@article{beucher2019uwgeodynamics,
+  title={UWGeodynamics: A teaching and research tool for numerical geodynamic modelling},
+  author={Beucher, Romain and Moresi, Louis and Giordani, Julian and Mansour, John and Sandiford, Dan and Farrington, Rebecca and Mondy, Luke and Mallard, Claire and Rey, Patrice and Duclaux, Guillaume and others},
[…]
