+++ paper/paper.md
@@ -44,10 +44,13 @@ Additionally, the code allows variations of boundary condition for the velocity 
 
 Although `Mandyoc` is the acronym for `MANtle DYnamics simulatOr Code`, it is designed to simulate Stokes flow type thermochemical convection taking different compositional layers into account, and it is also appropriate to simulate Earth's lithospheric dynamics in the geological timescale.
 
-Similar codes are available for scientific purposes, like ASPECT [@KHB12;@aspect2017;@Bangerth2021;@wolfgang_bangerth_2021_5131909;] and Underworld [@beucher2019uwgeodynamics].
+Similar codes are available for scientific purposes, like `ASPECT` [@KHB12;@aspect2017;@Bangerth2021;@wolfgang_bangerth_2021_5131909;] and `Underworld` [@moresi2007computational;@beucher2019uwgeodynamics].
 
 One advantage of `Mandyoc` is the possibility to create scenarios with velocity boundary conditions variable in space and time, allowing the user to simulate different tectonic pulses in the same model run.
 Additionally, the current version incorporates surface processes, imposing rates of erosion and sedimentation on the top of the free surface [@silva2022influence]. 
+%
+Recently, other thermomechanical codes available for the scientific community incorporated the interaction with surface processes [e.g. @beucher2020morphotectonic; @neuharth2022evolution] taking into account the simulation of fluvial and hillslope processes. In the present version of `Mandyoc`, only predefined erosion rate (variable in space and time) is possible. 
+%Other approaches for the surface processes are under development,
 
 Previous versions of the code were used to study the evolution of continental margins, showing the interaction of the continental lithosphere with the asthenospheric mantle [@sacek2017post;@salazar2021lateral].
 
