+++ paper/paper.bib
@@ -168,3 +168,35 @@
   pages={1136},
   year={2019}
 }
+
+@article{moresi2007computational,
+  title={Computational approaches to studying non-linear dynamics of the crust and mantle},
+  author={Moresi, L and Quenette, Steve and Lemiale, Vincent and Meriaux, C and Appelbe, B and M{\"u}hlhaus, H-B},
+  journal={Physics of the Earth and Planetary Interiors},
+  volume={163},
+  number={1-4},
+  pages={69--82},
+  year={2007},
+  publisher={Elsevier}
+}
+
+
+@article{neuharth2022evolution,
+  title={Evolution of rift systems and their fault networks in response to surface processes},
+  author={Neuharth, Derek and Brune, Sascha and Wrona, Thilo and Glerum, Anne and Braun, Jean and Yuan, Xiaoping},
+  journal={Tectonics},
+  pages={e2021TC007166},
+  year={2022},
+  publisher={Wiley Online Library}
+}
+
+@article{beucher2020morphotectonic,
+  title={Morphotectonic Evolution of Passive Margins Undergoing Active Surface Processes: Large-Scale Experiments Using Numerical Models},
+  author={Beucher, Romain and Huismans, RS},
+  journal={Geochemistry, Geophysics, Geosystems},
+  volume={21},
+  number={5},
+  pages={e2019GC008884},
+  year={2020},
+  publisher={Wiley Online Library}
+}
