### 2022-03-03 5108e13
Update in the model progress shown in stdout
