### 2022-03-03 d468041
Update in the model progress shown in stdout
