### 2022-03-03 ac7f55b
Model progress stdout updated
