### 2022-03-03 72e5988
Add docker images (#57)
* Add docker images and python requirements file needed to build images (to support running examples and tests)

* add workflow to build and publish docker image for releases

* fix gitignore for old mandyoc bin in root folder
