+++ env/requirements.txt
@@ -1,0 +1,4 @@
+numpy
+matplotlib
+pandas
+pytest
