### 2022-03-07 6bde2c5
Add a notebook to create the input file and plot the resultsfor the Crameri example.
Also, remove the old .py files that did this task.
