### 2022-03-08 c37241d
change tests structure; add more tests based on examples
