+++ tests/data/Crameri2012_case2/input/run.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+${MPIEXEC} -n 2 ${MANDYOC}
