+++ tests/conftest.py
@@ -1,0 +1,4 @@
+# conftest.py
+
+def pytest_addoption(parser):
+    parser.addoption("--scenario")
