+++ Makefile
@@ -47,9 +47,8 @@ install: $(MANDYOC)
 	install $< $(INSTALL_PATH)/mandyoc
 
 test:
-	@echo "Run MANDYOC test may take several minutes..."
-	cd test/testing_data/ ; mpirun -n 1 mandyoc
-	pytest -v test/testing_result.py
+	@echo "Running MANDYOC tests. It will take several minutes to complete all tests."
+	bash ./tests/run_tests.sh
 
 clear:
 	rm -f $(SRC)/*.o
