### 2022-03-08 7826d7d
improve installation instructions
