### 2022-03-08 390f396
add mandyoc version for Crameri benchmark
