### 2022-03-08 cb99d13
add scaling examples for vanKeken and Crameri examples (#67)
