### 2022-03-08 2692620
increase memory allocation for array used in for particles operations (#63)
