### 2022-03-08 f6f2059
add mandyoc version for vanKeken benchmark
