### 2022-03-08 10fa8e8
add mandyoc version for Crameri benchmark (#70)
* add mandyoc version for Crameri benchmark

* add mandyoc version for vanKeken benchmark
