### 2022-03-08 65dd0a8
Remove REDME ans fun files from Crameri example
