### 2022-03-08 be8335a
Create notebook for vanKeken examples and remove old files
The notebook creates and plots the results.
Remove: README file, run file, .py scripts and density plot.
