### 2022-03-08 2f7aae8
add missing petsc external packages necessary for direct solver
