### 2022-03-08 23b56db
Improve installation instructions (#69)
* improve installation instructions

* change word libraries to packages

* add missing petsc external packages necessary for direct solver
