+++ docs/files/installation.rst
@@ -8,32 +8,61 @@ and *Mandyoc*.
 	The following installation steps work for both Linux and macOS machines
 	**only** and no tests were made to install *Mandyoc* on Windows machines yet.
 
+.. _Dependencies:
+
 Dependencies
 ------------
 
-* PETSc_
+To build *Mandyoc*, the following requirements are needed:
+
+* PETSc_ (currently tested on version v3.15.5)
 * gcc
 * make
-* git
+* git (recommended, but not extrictly needded)
 
-**Optional** dependencies:
+If you do not already have a PETSc installation, you will also need:
 
 * gfortran
-* pytest
+
+Additionally, the following additional software is needed to run the examples
+that come with *Mandyoc*:
+
+* Python 3.5+
+
+With required python packages:
+
 * numpy
-* os
+* matplotlib
+* pandas
+
+To run the tests, some additional python packages are required:
+
+* pytest
+
+To build the documentation, further python packages are necessary:
+
+* sphinx
+* sphinx_rtd_theme
 
 PETSc Installation
 ------------------
 
-*Mandyoc* requires the `PETSc`_ library to run.
-The first step is to **download** the latest release of PETSc from `PETSc website`_
+*Mandyoc* requires the PETSc library to run.
+Check out the `PETSc installation`_ for details.
+
+.. note::
+
+	The following steps its a example installation with minimum requirements
+	to run *Mandyoc*.
+
+The first step is to **download** PETSc (v3.15.5) release from `PETSc website`_
 or **clone** the repository into your machine.
+*Mandyoc* might work with latest release of PETSc, but this is not guaranteed
+since new verions might introduce breaking changes.
 
-Choose the path to your PETSC installation and clone the repository::
+Clone the repository to your desired location::
 
-	cd /path/to/petsc
-	git clone -b release https://gitlab.com/petsc/petsc.git $HOME/petsc
+	git clone -b v3.15.5 https://gitlab.com/petsc/petsc.git
 
 Second, **configure the PETSc build** and set up the installation directory.
 
@@ -48,29 +77,27 @@ Second, **configure the PETSc build** and set up the installation directory.
 	  --with-cxx=g++ \
 	  --download-fblaslapack \
 	  --download-mpich \
-	  --download-hdf5 \
-	  --download-superlu_dist \
-	  --download-metis \
-	  --download-parmetis \
 	  --download-mumps \
 	  --download-scalapack \
-	  --download-cmake \
-	  COPTFLAGS='-O3 -march=native -mtune=native' \
-	  CXXOPTFLAGS='-O3 -march=native -mtune=native'
-
+	  --download-parmetis \
+	  --download-metis
 
 .. note::
 
-	If using ``gfortran`` optional dependency add the options
-	``--with-fc=gfortran`` and ``FOPTFLAGS='-O3 -march=native -mtune=native'``
-	to the PETSc build configuration above.
+	By default, *Mandyoc* uses direct solvers (LU and Cholesky) provided by `MUMPS`_.
+	This requires additional external packages. Refer to `PETSc documentation`_
+	for further information.
 
 .. note::
 
-	If you are build a development version of *Mandyoc* you can build
-	a **debug version** of PETSc by setting ``--with-debugging=1`` and removing
-	the ``COPTFLAGS``, ``CXXOPTFLAGS`` (and ``FOPTFLAGS``) flags.
+	If you want to build a development version of *Mandyoc*
+	its recommended to build a **debug version** of PETSc
+	by setting ``--with-debugging=1``.
 	In this case, you may set ``PETSC_ARCH=arch-label-debug``.
+
+.. note::
+
+	If you prefer *openmpi*, you need to swith ``--download-mpich`` to ``--download-openmpi``.
 
[…]
