### 2022-03-09 d39fa22
WIP - paper updates (#60)
* References to Aspect and Underworld

* New references + details about surface processes

* Corrections in the text

* Minor corrections in the manuscript

* Addition of line: Therefore,  is an alternative to preexistent softwares.

* Sentense about surface processes
