### 2022-03-09 b27d496
Fix the documentation about installation.
Add the information about the jupyter notebook and remove the
information about .py scripts
