### 2022-03-09 081fff7
Create the jupyter notebook for the examples (#58)
* Create a jupyter notebook to generate and plot the continental rift example

Create the notebook to replace the python scripts.

* Remove readme file.
This text is in the jupyter notebook now.

* Remove the run.sh script

This code is in the noteook now.

* Update the gitignore file.

Ignore: *.ipynb_*, input files generated by the notebook and the ouput
files generated after running Mandyoc.

* Improve notebook for continental_rif example.

Improve the plot in postprocessing section.

* Fix labels names un the temperature plot

* Add a note
[…]
