+++ examples/Crameri2012_case2/generate_input_files.py
@@ -1,215 +1,0 @@
-"""
-Generate the input files to run the Case 2 experiment presented by Crameri et al. (2012), which evaluates the sticky air method to obtain a numerical surface topography in geodynamic modelling.
-"""
-import numpy as np
-import matplotlib.pyplot as plt
-
-
-# total model horizontal extent (m)
-Lx = 2800 * 1.0e3
-
-# total model vertical extent (m)
-Lz = 850 * 1.0e3
-
-# number of points in horizontal direction
-Nx = 281
-
-# number of points in vertical direction
-Nz = 169
-
-# thickness of sticky air layer (m)
-H_sa = 150 * 1.0e3
-
-# thickness of lid (m)
-H_lid = 100 * 1.0e3
-
-# plume diameter (m)
-D_plume = 100 * 1.0e3
-
-# plume center (above the bottom) (m)
-H_plume = 300 * 1.0e3
-
-x = np.linspace(0, Lx, Nx)
-z = np.linspace(Lz, 0, Nz)
-X, Z = np.meshgrid(x, z)
-
-##############################################################################
-# Interfaces (bottom first)
-##############################################################################
-
-x_plume_left = int(((Lx - D_plume) / 2) / (Lx / (Nx - 1)))
-x_plume_right = int(((Lx + D_plume) / 2) / (Lx / (Nx - 1)))
-
-plume_bottom = -1 * np.ones(Nx) * (Lz - H_plume)
-plume_top = -1 * np.ones(Nx) * (Lz - H_plume)
-
-plume_bottom[x_plume_left:x_plume_right] = -1 * (Lz - H_plume) - np.sqrt(
-    np.power(D_plume / 2, 2) - np.power(x[x_plume_left:x_plume_right] - Lx / 2, 2)
-)
-plume_top[x_plume_left:x_plume_right] = -1 * (Lz - H_plume) + np.sqrt(
-    np.power(D_plume / 2, 2) - np.power(x[x_plume_left:x_plume_right] - Lx / 2, 2)
-)
-
-interfaces = {
-    "plume_bottom": plume_bottom,
-    "plume_top": plume_top,
-    "mantle_top": -1 * (np.ones(Nx) * (H_sa + H_lid)),
-    "lid_top": -1 * (np.ones(Nx) * (H_sa)),
-}
-
-with open("interfaces.txt", "w") as f:
-    layer_properties = f"""
-        C   1.0    0.1    1.0    100.0  0.01
-        rho 3300.0 3200.0 3300.0 3300.0 0.0
-        H   0.0    0.0    0.0    0.0    0.0
-        A   0.0    0.0    0.0    0.0    0.0
-        n   0.0    0.0    0.0    0.0    0.0
-        Q   0.0    0.0    0.0    0.0    0.0
-        V   0.0    0.0    0.0    0.0    0.0
-    """
-
-    for line in layer_properties.split("\n"):
-        line = line.strip()
-        if len(line):
-            f.write(" ".join(line.split()) + "\n")
-
-    # layer interfaces
-    data = np.array(tuple(interfaces.values())).T
-    np.savetxt(f, data, fmt="%f")
-
-##############################################################################
-# Plot interfaces
-##############################################################################
-fig, ax = plt.subplots(figsize=(16, 8))
-
-for label, layer in interfaces.items():
-    ax.plot(x / 1.0e3, layer / 1.0e3, label=f"{label}")
-
-ax.set_xticks(np.arange(0, Lx / 1.0e3 + 1, 100))
-ax.set_yticks(np.arange(-Lz / 1.0e3, 0 + 1, 50))
-
-ax.set_xlim([0, Lx / 1.0e3])
[…]
