### 2022-03-09 b8b23e7
add section about docker image
