### 2022-03-09 68d831a
Fix the documentation about installation. (#71)
Add the information about the jupyter notebook and remove the
information about .py scripts
