+++ examples/vanKeken1997_case1a/create_interface_file.py
@@ -1,160 +1,0 @@
-"""
-Create the input file to simulate the an Keken et al. (1997) model
-"""
-import numpy as np
-import matplotlib.pyplot as plt
-
-Nx = 81
-Nz = 81
-
-Lx = 92.42
-Lz = 100.0
-
-xn = np.linspace(0, Lx, Nx)
-
-liton = -0.8 * Lz + 0.02 * Lz * np.cos(np.pi * xn / Lx)
-
-liton = liton * 1000
-
-
-# Create the interface file
-
-layer_properties = [
-    "C    1.0     1.0",
-    "rho -1000.   0.",
-    "H    0.0e-12 0.0e-12",
-    "A    0.0     0.0",
-    "n    0.0     0.0",
-    "Q    0.0     0.0",
-    "V    0.0     0.0",
-]
-
-with open("interfaces.txt", "w") as f:
-    for line in layer_properties:
-        line = line.strip()
-        if len(line):
-            f.write(" ".join(line.split()) + "\n")
-
-    for line in np.arange(Nx):
-        f.write("%lf\n" % (liton[line]))
-
-
-##############################################################################
-# Parameters file
-##############################################################################
-params = f"""
-# Geometry
-nx                                  = {Nx}          # n. of nodes in the horizontal direction
-nz                                  = {Nz}          # n. of nodes in the vertical direction
-lx                                  = 91420.0       # extent in the horizontal direction
-lz                                  = 100000.       # extent in the vertical direction
-particles_per_element_x             = 0             # default is 0
-particles_per_element_z             = 0             # default is 0
-
-# Simulation options
-multigrid                           = 1             # ok -> soon to be on the command line only
-solver                              = direct        # default is direct [direct/iterative]
-denok                               = 1.0e-15       # default is 1.0e-4
-particles_per_element               = 1100          # default is 81
-particles_perturb_factor            = 0.0           # default is 0.5 [values are between 0 and 1]
-rtol                                = 1.0e-5        # the absolute size of the residual norm (relevant only for iterative methods), default is 1.0E-5
-RK4                                 = Euler         # default is Euler [Euler/Runge-Kutta]
-Xi_min                              = 1.0e-14       # default is 1.0e-14
-random_initial_strain               = 0.0           # default is 0.0
-pressure_const                      = -1.0          # default is -1.0 (not used)
-initial_dynamic_range               = False         # default is False [True/False]
-periodic_boundary                   = False         # default is False [True/False]
-high_kappa_in_asthenosphere         = False         # default is False [True/False]
-K_fluvial                           = 2.0e-7        # default is 2.0e-7
-m_fluvial                           = 1.0           # default is 1.0
-sea_level                           = 0.0           # default is 0.0
-basal_heat                          = -1.0          # default is -1.0
-
-# Surface processes
-sp_surface_tracking                 = False         # default is False [True/False]
-sp_surface_processes                = False         # default is False [True/False]
-sp_dt                               = 0             # default is 0
-sp_d_c                              = 0             # default is 0
-plot_sediment                       = False         # default is False [True/False]
-a2l                                 = True          # default is True [True/False]
-
-free_surface_stab                   = True          # default is True [True/False]
-theta_FSSA                          = 0.5           # default is 0.5 (only relevant when free_surface_stab = True)
-
-# Time constrains
-step_max                            = 4000          # maximum time-step of the simulation
-time_max                            = 140000.0e6    # maximum time of the simulation [years]
-dt_max                              = 1000.0e6      # maximum time between steps of the simulation [years]
-step_print                          = 20            # make file every <step_print>
-sub_division_time_step              = 1.0           # default is 1.0
-initial_print_step                  = 0             # default is 0
-initial_print_max_time              = 1.0e6         # default is 1.0e6 [years]
-
[…]
