### 2022-03-09 c73fcc5
removed double kaus from the reference file and fixed tabulation error in the input file
