+++ docs/files/refs.bib
@@ -162,7 +162,7 @@
   publisher={Cambridge University Press}
 }
 
-@article{druckerprager1952,
+@article{drucker-prager1952,
   title={Soil mechanics and plastic analysis or limit design},
   author={Drucker, Daniel Charles and Prager, William},
   journal={Quarterly of applied mathematics},
