### 2022-03-09 5fcf397
WIP Documentation changes (#56)
* Changed incorrect image; changed param.txt file in the documentation; changed van Keken's Python script to create proper param.txt file

* Changed word

Co-authored-by: victorsacek <sacek@usp.br>

* Removed unused line

Co-authored-by: victorsacek <sacek@usp.br>

* Changed Python code to print chenged word

Co-authored-by: victorsacek <sacek@usp.br>

* Removed line and changed word

Co-authored-by: victorsacek <sacek@usp.br>

* Removed line and changed word in the param.txt file

Co-authored-by: victorsacek <sacek@usp.br>

* Small correction and description wa
[…]
