+++ docs/files/src/scale_bcv.txt
@@ -1,0 +1,4 @@
+3 
+10.0     0.5 
+25.0    -1.0
+50.0    -2.0
