+++ docs/files/parameter-file.rst
@@ -3,59 +3,416 @@
 Parameter file 
 ==============
 
+.. role:: raw-html(raw)
+    :format: html
+
 The parameter file ``param.txt`` contains the information that is necessary for the simulation to run. 
 
 #. Geometry
 
-    * Number of elements in each direction 
-    * The extension of the model 
-
-#. Time constrains
-
-    * The time-step of the simulation
-    * Maximum time-step of the simulation
-    * Maximum time of the simulation (in seconds)
-    * Maximum time between steps of the simulation (in seconds)
-
-#. Viscosity values
-    
-    * The reference viscosity
-    * The minimum viscosity
-    * The maximum viscosity
-
-#. Physical parameters    
-
-    * Gravity acceleration
-    * Heat capacity
-    * Coefficient of thermal conductivity 
-    * Coefficient of thermal expansion
-
-#. Boundary conditions
-
-    * Temperature boundary condition at all borders
-    * Tangential velocity boundary conditions at all borders
-    * Normal velocity boundary conditions at all borders
-
-#. Rheology and temperature model numbers
-
-#. External initial temperature from ASCII file
-
-#. External initial interfaces from ASCII file
-
-#. External initial velocity files from ASCII file
-
-#. Type of solver
-
-    * Direct
-    * Iteractive 
+    * nx
+        default: :raw-html:`<br />` 
+        type: integer :raw-html:`<br />` 
+        unit: :raw-html:`<br />` 
+        definition: number of nodes in the horizontal direction
+    * nz
+        default: :raw-html:`<br />` 
+        type: integer :raw-html:`<br />` 
+        unit: :raw-html:`<br />` 
+        definition: number of nodes in the vertical direction
+    * lx                                  
+        default: :raw-html:`<br />` 
+        type: real :raw-html:`<br />` 
+        unit: m :raw-html:`<br />` 
+        definition: extent in the horizontal direction
+    * lz                                  
+        default: :raw-html:`<br />`
+        type: real :raw-html:`<br />`
+        unit: m :raw-html:`<br />`
+        definition: extent in the vertical direction
 
 #. Simulation options
 
-    * denok
-    * rtol
+    * solver                              
+        default: direct :raw-html:`<br />`
+        type: direct/iterative :raw-html:`<br />`
+        unit: :raw-html:`<br />`
+        definition: set the solver to be direct or iterative
+    * denok                               
+        default: 1.0e-4 :raw-html:`<br />`
+        type: real number :raw-html:`<br />`
+        unit: :raw-html:`<br />`
+        definition: tolerance criterion for the Uzawa's scheme
+    * rtol                                
+        default: 1.0e-5 :raw-html:`<br />`
+        type: real number :raw-html:`<br />`
+        unit: :raw-html:`<br />`
+        definition: the absolute size of the residual norm (relevant only for iterative methods)
+    * RK4                                 
+        default: Euler  :raw-html:`<br />`
+        type: Euler/Runge-Kutta :raw-html:`<br />`
+        unit: :raw-html:`<br />`
+        definition: particles advection method
+    * Xi_min                              
+        default: 1.0e-14        :raw-html:`<br />`
+        type: real number :raw-html:`<br />`
+        unit: :raw-html:`<br />`
+        definition: tolerance criterion for the convergence of the non-linear flow
+    * random_initial_strain               
+        default: 0.0            :raw-html:`<br />`
+        type: real number :raw-html:`<br />`
+        unit: :raw-html:`<br />`
+        definition: non-dimensional value for the initial strain perturbation for the entire domain
+    * pressure_const                      
+        default: -1.0 (i.e. not used) :raw-html:`<br />`
+        type: real number :raw-html:`<br />`
+        unit: Pa :raw-html:`<br />`
[…]
