+++ docs/files/refs.bib
@@ -155,6 +155,13 @@
   publisher={Elsevier Amsterdam}
 }
 
+@book{gerya2019,
+  title={Introduction to numerical geodynamic modelling},
+  author={Gerya, Taras},
+  year={2019},
+  publisher={Cambridge University Press}
+}
+
 @article{drucker-prager1952,
   title={Soil mechanics and plastic analysis or limit design},
   author={Drucker, Daniel Charles and Prager, William},

[…]
