### 2022-03-10 bcab8df
refactor tests; change to only first step
