+++ tests/conftest.py
@@ -1,4 +1,0 @@
-# conftest.py
-
-def pytest_addoption(parser):
-    parser.addoption("--scenario")
