+++ tests/data/Crameri2012_case2/input/run.sh
@@ -1,3 +1,6 @@
 #!/bin/sh
 
+# exit when any command fails
+set -e
+
 ${MPIEXEC} -n 2 ${MANDYOC}
