### 2022-03-10 2cb6d85
add section about docker image (#72)
