### 2022-03-10 a480cb4
fix copy of input files (remove output folder before copy)
