### 2022-03-11 145424b
Add automated tests (#68)
* change tests structure; add more tests based on examples

* add tests workflow (github action)

* refactor tests; change to only first step

* change to step_max = 1

* fix copy of input files (remove output folder before copy)

* set particles_perturb_factor to 0 for tests

* update Crameri expected data

* update continental_rift expected data

* update expected data

* update expected data for Crameri; change to run Crameri on 1 processor

* add absolute tolerance parameter

* update continental rift test setup and expected data

* add skip for some tests

* upda
[…]
