+++ tests/data/continental_rift/input/run.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+# exit when any command fails
+set -e
+
+${MPIEXEC} -n 2 ${MANDYOC} -seed 2 -strain_seed 1.0
