+++ tests/data/Crameri2012_case2/input/run.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+# exit when any command fails
+set -e
+
+${MPIEXEC} -n 1 ${MANDYOC}
