+++ .github/workflows/tests.yml
@@ -1,0 +1,36 @@
+name: tests
+
+# Controls when the workflow will run
+on:
+  # Allows run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+env:
+  TEST_TAG: ggciag/mandyoc:test
+
+jobs:
+  # Build the docker image and run tests
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+
+      - name: Build and export to Docker
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          file: docker/mandyoc.Dockerfile
+          load: true
+          tags: ${{ env.TEST_TAG }}
+
+      - name: Tests
+        run: |
+          docker run --rm ${{ env.TEST_TAG }} bash -c 'make test'
