### 2022-03-11 98db8c0
fix path fo mandyoc Dockerfile (#74)
