+++ .github/workflows/tests.yml
@@ -27,7 +27,7 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           context: .
-          file: docker/mandyoc.Dockerfile
+          file: docker/mandyoc/Dockerfile
           load: true
           tags: ${{ env.TEST_TAG }}
 
