### 2022-03-11 #4070 @rafaelmds
@rbeucher @psanan

We published a new release of [Mandyoc v.0.1.5](https://github.com/ggciag/mandyoc/releases/tag/v0.1.5) addressing most of the review issues.

We kept issues open, since we made comments on each of them.
