### 2022-03-16 f323e56
Add the information about the jupyter notebook and remove the (#77)
information about README files in benchmarks section
