### 2022-03-17 77f4b28
Documentation: information about the algorithm used in the non-linear flow
