+++ docs/files/refs.bib
@@ -226,3 +226,14 @@
   year={2003},
   publisher={Elsevier}
 }
+
+@article{thieulot2014,
+  title={ELEFANT: a user-friendly multipurpose geodynamics code},
+  author={Thieulot, C},
+  journal={Solid Earth Discussions},
+  volume={6},
+  number={2},
+  pages={1949--2096},
+  year={2014},
+  publisher={Copernicus GmbH}
+}
