+++ docs/files/refs.bib
@@ -227,13 +227,3 @@
   publisher={Elsevier}
 }
 
-@article{thieulot2014,
-  title={ELEFANT: a user-friendly multipurpose geodynamics code},
-  author={Thieulot, C},
-  journal={Solid Earth Discussions},
-  volume={6},
-  number={2},
-  pages={1949--2096},
-  year={2014},
-  publisher={Copernicus GmbH}
-}
