### 2022-03-17 d855d5b
Documentation: information about the algorithm used in the non-linear flow (#81)
* Documentation: information about the algorithm used in the non-linear flow

* Update docs/files/implementation.rst

* Update docs/files/refs.bib
