### 2022-03-17 c2b2a2d
Explicit citation that Mandyoc is a 2D code in the paper (#82)
