### 2022-03-18 #4070 @psanan
@jedbrown @victorsacek I've checked off everything on my review list - looks ready from my perspective! Thanks again for the submission and the code - I particularly appreciate the attention paid to thorough documentation, including easy-to-follow, well-discussed, and reproducible examples covering both a benchmark case and a more realistic model.
