### 2022-03-21 f46793f
Makefile: fix quoting of $MPIEXEC, which may contain arguments
With Open MPI, PETSc sets MPIEXEC = /path/to/mpiexec --oversubscribe
so that testing can use more than the number of physical cores. This
prevents "make test" from crashing.
