### 2022-03-21 #4070 @jedbrown
Thanks for your diligent reviews @psanan @rbeucher.

@victorsacek Once https://github.com/ggciag/mandyoc/pull/83 is merged (or you've decided to revise otherwise to address these points), please tag a release ([annotated tags](https://git-scm.com/book/en/v2/Git-Basics-Tagging) preferred) and archive your release on Zenodo or similar, then report the DOI back here.
