### 2022-03-21 44b7b06
JOSS edits (#83)
* Makefile: fix quoting of $MPIEXEC, which may contain arguments

With Open MPI, PETSc sets MPIEXEC = /path/to/mpiexec --oversubscribe
so that testing can use more than the number of physical cores. This
prevents "make test" from crashing.

* paper bib: fix missing DOI and redundant websites

* paper: light copy editing
