+++ paper/paper.bib
@@ -38,8 +38,7 @@
     and Richard Tran Mills and Todd Munson and Karl Rupp and Patrick Sanan
     and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang},
   title =  {{PETS}c {W}eb page},
-  url =    {https://www.mcs.anl.gov/petsc},
-  howpublished = {\url{https://www.mcs.anl.gov/petsc}},
+  url =    {https://petsc.org},
   year = {2021}
 }
 
@@ -52,8 +51,7 @@
   title  = {{PETS}c Users Manual},
   institution = {Argonne National Laboratory},
   year   = 2021,
-  number = {ANL-95/11 - Revision 3.15},
-  url    = {https://www.mcs.anl.gov/petsc}
+  number = {ANL-95/11 - Revision 3.15}
 }
 
 @InProceedings{petsc-efficient,
@@ -96,7 +94,8 @@
   volume={41},
   pages={e2021TC006808},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2021TC006808}
 }
 
 @article{van1997comparison,
@@ -121,7 +120,7 @@
 	year="2021", 
 	publisher="Zenodo", 
 	optkeywords="ASPECT", 
-	doi="http://doi.org/10.5281/zenodo.5131909", 
+	doi="10.5281/zenodo.5131909", 
 	opturl="https://doi.org/10.5281/zenodo.5131909"}
 
 @Article{Bangerth2021, 
@@ -129,7 +128,7 @@
 	title="ASPECT: Advanced Solver for Problems in Earth{\textquoteright}s ConvecTion, User Manual", 
 	year="2021", 
 	optkeywords="ASPET", 
-	doi="http://doi.org/10.6084/m9.figshare.4865333.v8", 
+	doi="10.6084/m9.figshare.4865333.v8", 
 	opturl="https://figshare.com/articles/journal_contribution/ASPECT_Advanced_Solver_for_Problems_in_Earth_s_ConvecTion_User_Manual/4865333"}
 
 @Article{KHB12, 
@@ -142,7 +141,7 @@
 	pages="12-29", 
 	optkeywords="ASPECT", 
 	issn="0956540X", 
-	doi="http://doi.org/10.1111/j.1365-246X.2012.05609.x", 
+	doi="10.1111/j.1365-246X.2012.05609.x", 
 	opturl="http://gji.oxfordjournals.org/cgi/doi/10.1111/j.1365-246X.2012.05609.x"}
 
 @Article{aspect2017, 
@@ -155,7 +154,7 @@
 	pages="833-851", 
 	optkeywords="ASPECT", 
 	issn="0956-540X", 
-	doi="http://doi.org/10.1093/gji/ggx195", 
+	doi="10.1093/gji/ggx195", 
 	opturl="https://academic.oup.com/gji/article-lookup/doi/10.1093/gji/ggx195"}
 
 
@@ -166,7 +165,8 @@
   volume={4},
   number={36},
   pages={1136},
-  year={2019}
+  year={2019},
+  doi={10.21105/joss.01136}
 }
 
 @article{moresi2007computational,
@@ -177,7 +177,8 @@
   number={1-4},
   pages={69--82},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.pepi.2007.06.009}
 }
 
 
@@ -187,7 +188,8 @@
   journal={Tectonics},
   pages={e2021TC007166},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2021TC007166}
 }
 
[…]
