### 2022-03-21 #4070 @victorsacek
Dear @jedbrown, @rbeucher, and @psanan

Thank you very much for the detailed revision of our code and documentation.
Zenodo archived the version v0.1.5.1 of the Manydoc code:
10.5281/zenodo.6373350
