### 2022-03-23 #4070 @victorsacek
Dear @jedbrown

We observed that the tag for the last release is an invalid semver. We chose four numbers (v0.1.5.1) instead of three.
The correct version must be v0.1.6.
What is the best option to correct this mistake?

Thank you for your attention.
