### 2022-03-24 01c46fa
add function to parse command line options
