+++ src/main.cpp
@@ -11,7 +11,7 @@ static char help[] = "\n\nMANDYOC: MANtle DYnamics simulatOr Code\n\n"\
 /* Victor Sacek                           */
 /* Jamison F. Assuncao                    */
 /* Agustina Pesce                         */
-/* Rafael M. da Silva                     */ 
+/* Rafael M. da Silva                     */
 /* [Contributed by Dave May]              */
 
 #include <petscksp.h>
@@ -48,17 +48,16 @@ PetscErrorCode sp_interpolate_surface_particles_to_vec();
 PetscErrorCode evaluate_surface_processes();
 PetscErrorCode sp_write_surface_vec(PetscInt i);
 PetscErrorCode sp_destroy();
-PetscErrorCode load_topo_var(int rank);
 PetscErrorCode rescalePrecipitation(double tempo);
+PetscErrorCode parse_options(int rank);
 
 int main(int argc,char **args)
 {
 	PetscErrorCode ierr;
 	char prefix[PETSC_MAX_PATH_LEN];
-	PetscInt Px,Pz;
 
 	ierr = PetscInitialize(&argc,&args,(char*)0,help);CHKERRQ(ierr);
-	
+
 	PetscPrintf(PETSC_COMM_WORLD,"           __  __              _   _   _____   __     __   ____     _____ \n");
 	PetscPrintf(PETSC_COMM_WORLD,"          |  \\/  |     /\\     | \\ | | |  __ \\  \\ \\   / /  / __ \\   / ____|\n");
 	PetscPrintf(PETSC_COMM_WORLD,"          | \\  / |    /  \\    |  \\| | | |  | |  \\ \\_/ /  | |  | | | |     \n");
@@ -66,7 +65,7 @@ int main(int argc,char **args)
 	PetscPrintf(PETSC_COMM_WORLD,"          | |  | |  / ____ \\  | |\\  | | |__| |    | |    | |__| | | |____ \n");
 	PetscPrintf(PETSC_COMM_WORLD,"          |_|  |_| /_/    \\_\\ |_| \\_| |_____/     |_|     \\____/   \\_____|\n");
 	PetscPrintf(PETSC_COMM_WORLD,"                                                                          \n");
-																	
+
 	PetscPrintf(PETSC_COMM_WORLD,"===================================================================================\n");
 	PetscPrintf(PETSC_COMM_WORLD,"=   MANDYOC: MANtle DYnamics simulatOr Code.\n");
 	PetscPrintf(PETSC_COMM_WORLD,"===================================================================================\n");
@@ -86,71 +85,17 @@ int main(int argc,char **args)
 	MPI_Comm_rank(PETSC_COMM_WORLD,&rank);
 
 	seed = rank;
-	
+
 	// Read ASCII files
 	reader(rank, "param.txt");
+
+	// Parse command line options
+	parse_options(rank);
 
 	PetscLogDouble Tempo1,Tempo2;
 	PetscTime(&Tempo1);
 	char variable_name[100];
 
-	Px = Pz = PETSC_DECIDE;
-	ierr = PetscOptionsGetInt(NULL,NULL,"-Px",&Px,NULL);CHKERRQ(ierr);
-	Pz = Px;
-	ierr = PetscOptionsGetInt(NULL,NULL,"-Pz",&Pz,NULL);CHKERRQ(ierr);
-
-	if (n_interfaces>0 && interfaces_from_ascii==1){
[…]
