+++ src/reader.cpp
@@ -7,9 +7,10 @@ PetscBool check_a_b_bool(char tkn_w[], char tkn_v[], const char str_a[], const c
 void ErrorInterfaces();
 
 // Parameter file variables
-extern long Nx, Nz;
+extern int dimensions;
+extern long Nx, Ny, Nz;
 extern long layers;
-extern double Lx, depth;
+extern double Lx, Ly, depth;
 extern int ContMult;
 extern long stepMAX;
 extern double timeMAX;
@@ -144,18 +145,18 @@ PetscErrorCode reader(int rank, const char fName[]){
 	char *tkn_w, *tkn_v;
 	char line[size];
 	// char *mandatoryParam = ["nx", "nz"];
-	
+
 	if (rank==0){
 		nline = 0;
 		FILE *f_parameters;
-		
+
 		f_parameters = fopen(fName,"r");
 		if (f_parameters == NULL) {PetscPrintf(PETSC_COMM_WORLD, "ERROR. The <%s> file was NOT FOUND.\n", fName); exit(1);}
 		while (!feof(f_parameters))
 		{
 			// Increment line number nline
 			nline += 1;
-			
+
 			// Read each line from the parameters f_parameters, trim undesireble characters
 			// and split line in thk_w[] and thk_v[].
 			nread = fgets(line, size, f_parameters);
@@ -164,12 +165,15 @@ PetscErrorCode reader(int rank, const char fName[]){
 			if (tkn_w[0] == '#') continue;
 			tkn_v 	= strtok(NULL, " \t=#\n");
 			// PetscPrintf(PETSC_COMM_WORLD, "(%s)(%s)\n", tkn_w, tkn_v);
-			
+
 			// Store and check every parameter.
 			// Parameters with values
-			if (strcmp(tkn_w, "nx") == 0) {Nx = atol(tkn_v);}
+			if (strcmp(tkn_w, "dimensions") == 0) {dimensions = atoi(tkn_v);}
+			else if (strcmp(tkn_w, "nx") == 0) {Nx = atol(tkn_v);}
+			else if (strcmp(tkn_w, "ny") == 0) {Ny = atol(tkn_v);}
 			else if (strcmp(tkn_w, "nz") == 0) {Nz = atol(tkn_v);}
 			else if (strcmp(tkn_w, "lx") == 0) {Lx = atof(tkn_v);}
+			else if (strcmp(tkn_w, "ly") == 0) {Ly = atof(tkn_v);}
 			else if (strcmp(tkn_w, "lz") == 0) {depth = atof(tkn_v);}
 			else if (strcmp(tkn_w, "multigrid") == 0) {ContMult = atoi(tkn_v);}
 			else if (strcmp(tkn_w, "step_max") == 0) {stepMAX = atol(tkn_v);}
@@ -211,7 +215,7 @@ PetscErrorCode reader(int rank, const char fName[]){
 			else if (strcmp(tkn_w, "basal_heat") == 0) {basal_heat = atof(tkn_v);}
 			else if (strcmp(tkn_w, "sp_dt") == 0) {sp_dt = atof(tkn_v);}
 			else if (strcmp(tkn_w, "sp_d_c") == 0) {sp_d_c = atof(tkn_v);}
-			
+
 			// Boolean parameters
 			else if (strcmp(tkn_w, "geoq") == 0) {geoq_on = check_a_b(tkn_w, tkn_v, "on", "off");}
[…]
