### 2022-03-24 3b46361
remove 2d and 3d suffixes from function names
