+++ src/DMT.cpp
@@ -94,21 +94,21 @@ extern double kappa;
 extern double RHOM;
 extern double c_heat_capacity;
 
-PetscErrorCode create_thermal_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz)
+PetscErrorCode create_thermal(PetscInt mx, PetscInt mz, PetscInt Px, PetscInt Pz)
 {
 
 	PetscInt       dof,stencil_width;
 	PetscErrorCode ierr;
-	
-	
-	
+
+
+
 	int rank;
 	MPI_Comm_rank(PETSC_COMM_WORLD,&rank);
-	
+
 	PetscLogDouble Tempo1p,Tempo2p;
-	
+
 	PetscTime(&Tempo1p);
-	
+
 	dof           = 1;
 	if (periodic_boundary==0)	{
 		stencil_width = 1;
@@ -123,36 +123,36 @@ PetscErrorCode create_thermal_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz
 
 	ierr = DMSetFromOptions(da_Thermal);CHKERRQ(ierr);
 	ierr = DMSetUp(da_Thermal);CHKERRQ(ierr);
-	
+
 	ierr = DMDASetFieldName(da_Thermal,0,"T");CHKERRQ(ierr);
 
-	
-	
-	
+
+
+
 	ierr = PetscCalloc1(GaussQuad*T_NE,&NT); CHKERRQ(ierr);
 	ierr = PetscCalloc1(GaussQuad*T_NE,&NT_x); CHKERRQ(ierr);
 	ierr = PetscCalloc1(GaussQuad*T_NE,&NT_z); CHKERRQ(ierr);
-	
+
 	ierr = PetscCalloc1(T_NE*T_NE,&TKe); CHKERRQ(ierr);
 	ierr = PetscCalloc1(T_NE*T_NE,&TCe); CHKERRQ(ierr);
 	ierr = PetscCalloc1(T_NE*T_NE,&TCe_fut); CHKERRQ(ierr);
 	ierr = PetscCalloc1(T_NE*T_NE,&TMe); CHKERRQ(ierr);
 	ierr = PetscCalloc1(T_NE*T_NE,&Ttotal); CHKERRQ(ierr);
 	ierr = PetscCalloc1(T_NE*T_NE,&Ttotal_b); CHKERRQ(ierr);
-	
+
 	ierr = PetscCalloc1(T_NE,&TFe); CHKERRQ(ierr);
-	
+
 	ierr = PetscCalloc1(T_NE,&T_vec_aux_ele); CHKERRQ(ierr);
 	ierr = PetscCalloc1(T_NE,&T_vec_aux_ele_final); CHKERRQ(ierr);
-	
+
 	ierr = PetscCalloc1(V_GT,&v_vec_aux_ele); CHKERRQ(ierr);
-	
+
 	montaKeThermal_general(TKe,TMe,TFe);
-	
-	
+
+
 	ierr = DMDASetUniformCoordinates(da_Thermal,0.0,Lx,-depth,0.0,0.0,0.0);CHKERRQ(ierr);
-	
-		
+
+
 	/* Generate a matrix with the correct non-zero pattern of type AIJ. This will work in parallel and serial */
 	ierr = DMSetMatType(da_Thermal,MATAIJ);CHKERRQ(ierr);
 	ierr = DMCreateMatrix(da_Thermal,&TA);CHKERRQ(ierr);
@@ -163,19 +163,19 @@ PetscErrorCode create_thermal_2d(PetscInt mx,PetscInt mz,PetscInt Px,PetscInt Pz
 
 	ierr = DMCreateGlobalVector(da_Thermal,&Pressure_aux);CHKERRQ(ierr);
 
-	
[…]
