### 2022-03-24 9672021
add new parameters (Ny, Ly) to functions
