### 2022-03-24 2cecce7
change Calc_dt_calor to return the value (instead of updating global variable)
