### 2022-03-24 a028546
update Calc_dt_calor to 2D or 3D cases
