### 2022-03-25 95e7be3
update create_thermal to 2D/3D modes
