+++ src/DMT.cpp
@@ -27,6 +27,7 @@ extern double dz_const;
 
 extern PetscReal *NT;
 extern PetscReal *NT_x;
+extern PetscReal *NT_y;
 extern PetscReal *NT_z;
 
 extern long GaussQuad;
@@ -94,10 +95,11 @@ extern double kappa;
 extern double RHOM;
 extern double c_heat_capacity;
 
-PetscErrorCode create_thermal(PetscInt mx, PetscInt my, PetscInt mz, PetscInt Px, PetscInt Py, PetscInt Pz)
+PetscErrorCode create_thermal(int dimensions, PetscInt mx, PetscInt my, PetscInt mz, PetscInt Px, PetscInt Py, PetscInt Pz)
 {
 
 	PetscInt       dof,stencil_width;
+	DMBoundaryType boundary_type;
 	PetscErrorCode ierr;
 
 
@@ -109,16 +111,27 @@ PetscErrorCode create_thermal(PetscInt mx, PetscInt my, PetscInt mz, PetscInt Px
 
 	PetscTime(&Tempo1p);
 
-	dof           = 1;
-	if (periodic_boundary==0)	{
+	dof = 1;
+
+	if (periodic_boundary == 0) {
 		stencil_width = 1;
-		ierr          = DMDACreate2d(PETSC_COMM_WORLD,DM_BOUNDARY_NONE,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
-								 mx+1,mz+1,Px,Pz,dof,stencil_width,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
-	}
-	else {
+		boundary_type = DM_BOUNDARY_NONE;
+	} else {
 		stencil_width = 2;
-		ierr          = DMDACreate2d(PETSC_COMM_WORLD,DM_BOUNDARY_PERIODIC,DM_BOUNDARY_NONE,DMDA_STENCIL_BOX,
-								 mx+1,mz+1,Px,Pz,dof,stencil_width,NULL,NULL,&da_Thermal);CHKERRQ(ierr);
+		boundary_type = DM_BOUNDARY_PERIODIC;
+	}
+
+	if (dimensions == 2) {
+		ierr = DMDACreate2d(PETSC_COMM_WORLD, boundary_type, DM_BOUNDARY_NONE, DMDA_STENCIL_BOX,
+							mx+1, mz+1, Px, Pz, dof, stencil_width, NULL, NULL, &da_Thermal); CHKERRQ(ierr);
+	} else {
+		//
+		// NOTE: periodic_boundary is currently not used in 3D
+		//
+		ierr = PetscPrintf(PETSC_COMM_WORLD,"\n\nWARNING: periodic_boundary is currently not supported in 3D mode\n\n");
+
+		ierr = DMDACreate3d(PETSC_COMM_WORLD, DM_BOUNDARY_NONE, DM_BOUNDARY_NONE, DM_BOUNDARY_NONE, DMDA_STENCIL_BOX,
+							mx+1, my+1, mz+1, Px, Py, Pz, dof, stencil_width, NULL, NULL, NULL, &da_Thermal); CHKERRQ(ierr);
 	}
 
 	ierr = DMSetFromOptions(da_Thermal);CHKERRQ(ierr);
@@ -131,6 +144,9 @@ PetscErrorCode create_thermal(PetscInt mx, PetscInt my, PetscInt mz, PetscInt Px
 
[…]
