### 2022-03-25 b6b0b24
add functions for 3d mode (2d tests - first pass)
