+++ src/DMSwarm_2d.cpp
@@ -1,0 +1,503 @@
+#include <petscsf.h>
+#include <petscdm.h>
+#include <petscksp.h>
+#include <petscdmda.h>
+#include <petscdmshell.h>
+#include <petscdmswarm.h>
+#include <petsc/private/dmimpl.h>
+#include <petscmath.h>
+
+extern DM dmcell;
+
+extern DM dms;
+
+extern DM da_Veloc;
+
+extern long Nx;
+
+extern double dx_const;
+extern double dz_const;
+
+extern double Lx, depth;
+
+extern double H_lito;
+extern double escala_viscosidade;
+
+extern PetscInt particles_per_ele;
+extern PetscInt nx_ppe;
+extern PetscInt nz_ppe;
+
+extern double H_per_mass;
+
+extern double h_air;
+
+extern double RHOM;
+
+extern int n_interfaces;
+extern PetscScalar *interfaces;
+
+extern PetscScalar *inter_rho;
+extern PetscScalar *inter_geoq;
+extern PetscScalar *inter_H;
+
+extern PetscInt particles_add_remove;
+
+extern PetscReal particles_perturb_factor;
+
+extern PetscInt *ppp;
+extern PetscInt *p_remove;
+extern PetscInt *p_i;
+
+extern PetscReal *p_add_coor;
+extern PetscReal *p_add_r;
+extern PetscInt *p_add_i;
+extern PetscInt *p_add_layer;
+extern PetscReal *p_add_r_strain;
+extern PetscReal *p_add_r_strain_rate;
+
+extern unsigned int seed;
+
+extern PetscInt print_step_files;
+
+extern PetscInt *seed_layer;
+extern PetscInt seed_layer_size;
+extern PetscBool seed_layer_set;
+
+
+extern PetscReal *strain_seed_layer;
+extern PetscInt strain_seed_layer_size;
+extern PetscBool strain_seed_layer_set;
+
+extern PetscInt checkered;
+
+extern PetscBool plot_sediment;
+
+extern PetscReal random_initial_strain;
+
+extern PetscInt binary_output;
+
+
+
+PetscErrorCode _DMLocatePoints_DMDARegular_3d_2d_IS_2d(DM dm,Vec pos,IS *iscell)
+{
+
+	PetscInt p,n,bs,npoints,si,sk,milocal,mklocal,mx,mz;
+	DM dmregular;
+	PetscInt *cellidx;
+	PetscScalar *coor;
+	PetscReal dx,dz;
+	PetscErrorCode ierr;
+	PetscMPIInt rank;
+
+	ierr = MPI_Comm_rank(PETSC_COMM_WORLD,&rank);CHKERRQ(ierr);
+	ierr = VecGetLocalSize(pos,&n);CHKERRQ(ierr);
+	ierr = VecGetBlockSize(pos,&bs);CHKERRQ(ierr);
+	npoints = n/bs;
+
+	PetscMalloc1(npoints,&cellidx);
+
+	ierr = DMGetApplicationContext(dm,(void**)&dmregular);CHKERRQ(ierr);
+	ierr = DMDAGetCorners(dmregular,&si,&sk,NULL,&milocal,&mklocal,NULL);CHKERRQ(ierr);
+	ierr = DMDAGetInfo(dmregular,NULL,&mx,&mz,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL);CHKERRQ(ierr);
+
+
+
[…]
