+++ src/DM1_v.cpp
@@ -3,9 +3,11 @@
 #include <petscsys.h>
 #include <petsctime.h>
 
-PetscErrorCode DMDAGetElementCorners(DM da,PetscInt *sx,PetscInt *sz,PetscInt *mx,PetscInt *mz);
+PetscErrorCode DMDAGetElementCorners_2d(DM da,PetscInt *sx,PetscInt *sz,PetscInt *mx,PetscInt *mz);
+PetscErrorCode DMDAGetElementCorners_3d(DM da,PetscInt *sx,PetscInt *sy,PetscInt *sz,PetscInt *mx,PetscInt *my,PetscInt *mz);
 
-PetscReal montaKeVeloc_simplif(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
+PetscReal montaKeVeloc_simplif_2d(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
+PetscReal montaKeVeloc_simplif_3d(PetscReal *Ke,PetscReal *KeG, PetscReal *geoq_ele);
 
 PetscErrorCode write_veloc(int cont, PetscInt binary_out);
 PetscErrorCode write_veloc_cond(int cont, PetscInt binary_out);
@@ -15,7 +17,13 @@ PetscErrorCode ascii2bin(char *s1, char *s2);
 typedef struct {
 	PetscScalar u;
 	PetscScalar w;
-} Stokes;
+} Stokes2d;
+
+typedef struct {
+	PetscScalar u;
+	PetscScalar v;
+	PetscScalar w;
+} Stokes3d;
 
 extern PetscInt binary_output;
 
@@ -37,6 +45,7 @@ extern PetscReal *Ke_veloc_final;
 extern PetscReal *Ke_veloc_general;
 
 extern double dx_const;
+extern double dy_const;
 extern double dz_const;
 
 extern double depth;
@@ -84,6 +93,9 @@ extern Vec local_Precon;
 extern double visc_aux_MAX;
 extern double visc_aux_MIN;
 
+extern double e2_aux_MAX;
+extern double e2_aux_MIN;
+
 extern double seg_per_ano;
 
 extern Vec Veloc;
@@ -99,6 +111,8 @@ extern double depth;
 extern PetscInt veloc_extern;
 extern PetscInt i_veloc;
 
+extern PetscInt print_visc;
+
 extern int tcont;
 extern long print_step;
 
@@ -112,7 +126,7 @@ extern PetscReal theta_FSSA;
 
 extern PetscInt periodic_boundary;
 
-PetscErrorCode AssembleA_Veloc(Mat A,Mat AG,DM veloc_da, DM temper_da){
+PetscErrorCode AssembleA_Veloc_2d(Mat A,Mat AG,DM veloc_da, DM temper_da){
 
 	PetscErrorCode         ierr;
 
@@ -124,12 +138,12 @@ PetscErrorCode AssembleA_Veloc(Mat A,Mat AG,DM veloc_da, DM temper_da){
 
 	PetscScalar u[V_GT*V_GT],val_cond[1];
 
-	Stokes					**pc;
+	Stokes2d					**pc;
 
 	ierr = VecZeroEntries(local_Precon);CHKERRQ(ierr);
[…]
