### 2022-03-27 #4070 @jedbrown
@victorsacek You can either tag v0.1.6 and update Zenodo or we can work with v0.1.5.1 (JOSS doesn't require that you use semver, strictly or otherwise). Your choice, just let us know which you choose.
